==> analysis/scripts/05_subcluster.Rmd <==
---
title: "Subclustering"
author: "David Wu"
---

### Use the here package for robust relative paths
```{r}
library(here)
```

### Optionally convert this Rmd to R script 
```{r eval=FALSE, message=FALSE}
current_file <- rstudioapi::getActiveDocumentContext()$path
output_file <- stringr::str_replace(current_file, '.Rmd', '.R')
knitr::purl(current_file, output = output_file)
file.edit(output_file)
```

### Directories
```{r, message=FALSE}
output_dir <- here('analysis/output/05_subcluster') # analysis file output directory
data_dir <- here('data/derived/tildra') # data file output directory

dir.create(output_dir, showWarnings = FALSE)
dir.create(data_dir, showWarnings = FALSE)
```

### Loading the libraries
```{r, message=FALSE}
library(Seurat) 
library(tidyverse)
library(ggExtra)
library(SeuratWrappers)
library(wutilities) # devtools::install_github('symbiologist/wutilities')
library(tictoc)
library(harmony)
```
### Functions

```{r}
recluster <- function(seuratobj, 
                      assay = 'RNA',
                      reduction = 'harmony',
                      dims = 1:20,
                      seed = 42,
                      n.neighbors = 30,
                      n.epochs = 200,
                      min.dist = 0.3,
                      k.param = n.neighbors,
                      metric = 'cosine',
                      algorithm = 1,
                      resolution = c(0.1, 1.0, 2.0),
                      default_ident = 'RNA_snn_res.1') {
  
    ### RunUMAP, FindNeighbors, and FindClusters 
  
  seuratobj <- seuratobj %>% 
    FindNeighbors(reduction = reduction, 
                  dims = dims, 
                  assay = assay,
                  annoy.metric = metric,
                  k.param = k.param) %>% 
    FindClusters(random.seed = seed,
                 algorithm = algorithm,
                 resolution = resolution) %>% 
    RunUMAP(reduction = reduction, 
            dims = dims, 
            assay = assay, 
            seed.use = seed,
            n.neighbors = n.neighbors,
            n.epochs = n.epochs,
            min.dist = min.dist)
  
  ### Fix cluster factor levels
  clusterings <- colnames(seuratobj@meta.data) %>% str_subset('_res')
  
  for(i in clusterings) {
    clusters <- seuratobj@meta.data[[i]]
    seuratobj[[i]] <- factor(clusters, levels = levels(clusters) %>% as.numeric() %>% sort())
  }
  
  Idents(seuratobj) <- default_ident
  seuratobj$seurat_clusters <- seuratobj[[default_ident]]
  
  seuratobj
}
```

### Global seed
```{r}
seed <- 123
set.seed(seed)
```

### Import Seurat
```{r}
seuratobj <- read_rds(file.path(data_dir, 'seuratobj_clustered.rds'))
seuratobj
```

### Visualize
```{r}
p <- seurat_feature(seuratobj, facet_hide = TRUE)

p
```


```{r}
p <- seurat_feature(seuratobj, features = 'seurat_clusters', facets = 'seurat_clusters', label = FALSE)

p
```


### Major cell types
```{r}
p <- seurat_feature(seuratobj, 
               features = c('CD3D', #lymphocytes
                            'MS4A1', # b cells
                            'HLA-DRA', # APC
                            'LYZ', # APC
                            'TPSAB1', # mast cells
                            'COL6A2', # fibroblasts
                            'PECAM1', # endothelial cells
                            'MKI67'), # cycling
               nrow = 2, 
               rasterize_dpi = 600) 

ggsave(plot = p,
       filename = 'broad_markers.png',
       h = 4,
       w = 8,
       dpi = 600,
       path = output_dir)

p
```



#### Calculate markers from broad clusters
```{r}
broad_markers <- RunPrestoAll(seuratobj, min.pct = 0.25, only.pos = TRUE, min.diff.pct = 0.25) %>% 
  mutate(pct.diff = pct.1 - pct.2) %>% 
  arrange(-pct.diff) %>% 
  select(cluster, gene, everything()) %>% 
  as_tibble()

broad_markers %>% group_by(cluster) %>% slice_max(pct.diff, n = 5)
```

```{r}
broad_markers %>% group_by(cluster) %>% slice_max(pct.diff, n = 5)
```
```{r}
broad_markers %>% filter(gene %in% 'COL1A1')
```

### Subclusters
#### Major cell types
```{r}
mast_clusters <- broad_markers %>% filter(gene == 'TPSAB1') %>% pull(cluster) %>% unique()
stromal_clusters <- broad_markers %>% filter(gene %in% c('COL1A1', 'PECAM1')) %>% pull(cluster) %>% unique()
bcell_clusters <- broad_markers %>% filter(gene %in% c('MS4A1', 'CD79A', 'IGKC', 'JCHAIN')) %>% pull(cluster) %>% unique()
cycling_clusters <- broad_markers %>% filter(gene == 'MKI67') %>% pull(cluster) %>% unique()
apc_clusters <- broad_markers %>% filter(gene == 'HLA-DRA') %>% pull(cluster) %>% unique()
lymphocyte_clusters <- broad_markers %>% filter(gene %in% c('CD3D', 'CD4', 'CD8A')) %>% pull(cluster) %>% unique()

```

```{r}
celltypes <- seuratobj@meta.data %>% 
  select(id, seurat_clusters) %>% 
  mutate(celltype = case_when(seurat_clusters %in% cycling_clusters ~ 'Cycling',
                              seurat_clusters %in% bcell_clusters ~ 'B',
                              seurat_clusters %in% stromal_clusters ~ 'Stromal',
                              seurat_clusters %in% mast_clusters ~ 'Mast',
                              seurat_clusters %in% apc_clusters ~ 'APC',
                              seurat_clusters %in% lymphocyte_clusters ~ 'Lymphocyte'))


seuratobj$celltype <- celltypes$celltype
p <- seurat_feature(seuratobj, features = 'celltype', facet_hide = TRUE, legend_position = 'none', rasterize_dpi = 600)

ggsave(plot = p,
       filename = 'broad_celltypes.png',
       h = 5,
       w = 5,
       dpi = 600,
       path = output_dir)
p
```


```{r}
celltype_tally <- celltypes %>% 
  group_by(id, celltype) %>% 
  tally() %>% 
  pivot_wider(names_from = 'celltype',
              values_from = 'n',
              values_fill = 0,
              names_prefix = 'n_')

celltype_tally %>% write_tsv(here(output_dir, 'celltype_tally.tsv'))

celltype_tally
```


### Iterative subsclustering and re-harmonize
```{r}
batch <- c('treatment', 'patient', 'flow', 'mouse', 'chemistry') # global batch variables

apcs <- subset(seuratobj, subset = celltype == 'APC') 
lymphocytes <- subset(seuratobj, subset = celltype == 'Lymphocyte')

p1 <- seurat_feature(apcs, facet_hide = TRUE, title = 'APCs')
p2 <- seurat_feature(lymphocytes, facet_hide = TRUE, title = 'Lymphocytes')
p1 + p2
```

#### Reprocess subclusters
```{r}
apcs <- apcs %>% DietSeurat() %>% FindVariableFeatures(nfeatures = 3000) %>% ScaleData() %>% RunPCA() 

ElbowPlot(apcs, ndims = 50)
```

#### Select dimensions prior to harmony as all dimensions are used
```{r}
max_harmony_dims <- 20

apcs <- apcs %>% RunHarmony(group.by.vars = batch,
                            dims.use = 1:max_harmony_dims)

apcs %>% ElbowPlot(ndims = max_harmony_dims, reduction = 'harmony')
```

```{r}
lymphocytes <- lymphocytes %>% DietSeurat() %>% FindVariableFeatures(nfeatures = 3000) %>% ScaleData() %>% RunPCA() 

ElbowPlot(lymphocytes, ndims = 50)
```

```{r}
lymphocytes <- lymphocytes %>% RunHarmony(group.by.vars = batch,
                                          dims.use = 1:max_harmony_dims)

lymphocytes %>% ElbowPlot(ndims = max_harmony_dims, reduction = 'harmony')
```

### Cluster subgroups separately, then recombine objects for grouped analysis
#### Reclustering parameters
```{r}
dims <- 1:max_harmony_dims
```

```{r}
apcs <- apcs %>% recluster(dims = dims, seed = seed)
```

```{r}
p <- seurat_feature(apcs, facet_hide = TRUE, title = 'APC Subcluster', legend_position = 'none')

ggsave(plot = p,
       filename = 'umap_apc.png',
       h = 5,
       w = 5,
       path = output_dir)
p
```

```{r}
apcs@meta.data %>% 
  select(seurat_clusters, id) %>% 
  unique() %>% 
  add_count(seurat_clusters) %>% 
  arrange(n)
```

#### Lymphocytes
```{r}
lymphocytes <- lymphocytes %>% recluster(dims = dims, seed = seed)
```

```{r}
p <- seurat_feature(lymphocytes, facet_hide = TRUE)
p
```
```{r}
p <- seurat_feature(lymphocytes, facet_hide = TRUE, title = 'Lymphocyte Subcluster', legend_position = 'none')

ggsave(plot = p,
       filename = 'umap_lymphocyte.png',
       h = 5,
       w = 5,
       path = output_dir)
p
```

```{r}
lymphocytes@meta.data %>%
  select(seurat_clusters, id) %>%
  unique() %>%
  add_count(seurat_clusters) %>%
  arrange(n)
```

### Recombine and recluster
```{r}
recombined_cells <- celltypes %>% filter(celltype %in% c('APC', 'B', 'Lymphocyte', 'Mast')) %>% rownames()

subcluster <- subset(seuratobj, cells = recombined_cells) %>% DietSeurat()

combined_features <- union(apcs@assays$RNA@var.features, lymphocytes@assays$RNA@var.features)
VariableFeatures(subcluster) <- combined_features

subcluster <- subcluster %>% 
  ScaleData() %>% 
  RunPCA() 

subcluster %>% ElbowPlot(ndims = 50)
```
```{r}
max_harmony_dims <- 20
subcluster <- subcluster %>% RunHarmony(group.by.vars = batch,
                                        dims.use = 1:max_harmony_dims)

ElbowPlot(subcluster, ndims = max_harmony_dims, reduction = 'harmony')
```

### RunUMAP, FindNeighbors, and FindClusters 
```{r}
subcluster <- subcluster %>% recluster(dims = 1:max_harmony_dims,
                                       resolution = c(0.1, 1, 2),
                                       n.epochs = 500,
                                       default_ident = 'RNA_snn_res.2',
                                       seed = seed)

subcluster$global <- Idents(subcluster)
```

```{r}
subcluster %>% seurat_feature(feature = 'global')
```


### Any clusters dominated by specific patients?
```{r}
sample_cluster_tally <- subcluster@meta.data %>% 
  select(global, id, patient, sample) %>% 
  unique() %>% 
  add_count(global) %>% 
  arrange(n)

sample_cluster_tally
```

### Add cluster information from iterative subclustering 
```{r}
apc_clustering <- apcs@meta.data %>% 
  rownames_to_column() %>% 
  select(rowname, seurat_clusters) %>% 
  mutate(local = factor(paste0('A', seurat_clusters), levels = paste0('A', levels(seurat_clusters))))

lymphocyte_clustering <- lymphocytes@meta.data %>% 
  rownames_to_column() %>% 
  select(rowname, seurat_clusters) %>% 
  mutate(local = factor(paste0('L', seurat_clusters), levels = paste0('L', levels(seurat_clusters))))

other_clustering <- celltypes %>% 
  rownames_to_column() %>% 
  select(rowname, seurat_clusters, local = celltype) %>% 
  filter(local %in% c('B', 'Mast')) %>% 
  mutate(local = factor(local))

merged_clustering <- subcluster@meta.data %>% 
  rownames_to_column() %>% 
  select(rowname) %>% left_join(
    bind_rows(lymphocyte_clustering,
              apc_clustering,
              other_clustering) %>% 
      select(-seurat_clusters)) %>% 
  column_to_rownames()

merged_clustering
```
```{r}
subcluster <- AddMetaData(subcluster, merged_clustering)
```


```{r}
subcluster %>% seurat_feature(feature = 'local', facet_hide = TRUE, title = 'Lymphocyte and APC Clusters', legend_position = 'none')
```

```{r}
p3 <- subcluster %>% seurat_feature(facet_hide = TRUE, 
                                    feature = 'global',
                                    title = 'Global Clusters', legend_position = 'none', color_package = 'carto', color_palette = 'Bold', alpha = 0.2, label_size = 4,
                                    rasterize_dpi = 600)

ggsave(plot = p3,
       filename = 'umap_global_labeled.png',
       dpi = 600,
       h = 5,
       w = 5,
       path = output_dir)

p3
```
```{r}
p4 <- subcluster %>% 
  seurat_feature(facet_hide = TRUE, 
                 feature = 'global',
                 title = 'Global Clusters', 
                 legend_position = 'none', color_package = 'carto', color_palette = 'Bold', alpha = 0.2, label = FALSE, rasterize_dpi = 600)

ggsave(plot = p4,
       filename = 'umap_global.png',
       dpi = 600,
       h = 5,
       w = 5,
       path = output_dir)

p4
```

```{r}
p5 <- subcluster %>% 
  seurat_feature(facet_hide = TRUE, 
                 feature = 'local',
                 title = 'Local Clusters', 
                 legend_position = 'none', color_package = 'carto', color_palette = 'Bold', alpha = 0.2, label = FALSE, rasterize_dpi = 600)

ggsave(plot = p5,
       filename = 'umap_local.png',
       dpi = 600,
       h = 5,
       w = 5,
       path = output_dir)

p5
```

```{r}
p6 <- subcluster %>% seurat_feature(title = 'Local Clusters', 
                                    feature = 'local', 
                                    facet_hide = TRUE,
                                    legend_position = 'none', 
                                    color_package = 'carto', 
                                    color_palette = 'Bold', 
                                    alpha = 0.2, 
                                    label = TRUE)

ggsave(plot = p6,
       filename = 'umap_local_labeled.png',
       h = 12,
       w = 12,
       path = output_dir)

p6
```
```{r}
p7 <- subcluster %>% seurat_feature(title = 'Global Clusters', 
                                    feature = 'global',
                                    facets = 'global', 
                                    legend_position = 'none', color_package = 'carto', color_palette = 'Pastel', alpha = 0.2, label = FALSE)

ggsave(plot = p7,
       filename = 'umap_global_facets.png',
       h = 12,
       w = 12,
       path = output_dir)
```

```{r}
p8 <- subcluster %>% seurat_feature(title = 'Local Clusters', 
                                    feature = 'local',
                                    facets = 'local', 
                                    legend_position = 'none', color_package = 'carto', color_palette = 'Pastel', alpha = 0.2, label = FALSE)

ggsave(plot = p8,
       filename = 'umap_local_facets.png',
       h = 12,
       w = 12,
       path = output_dir)
```

## Export
```{r}
subcluster %>% write_rds(here(data_dir, 'seuratobj_subcluster.rds'))
```

## Session Info
```{r}
sessionInfo()
```
