==> analysis/scripts/09_il23_profile.Rmd <==
---
title: "DE analysis"
author: "David Wu"
output:
  pdf_document: dev: cairo_pdf
---

Set working directory to project directory
```{r}
library(here)
```


### Optionally convert this Rmd to R script 
```{r eval=FALSE, message=FALSE}
current_file <- rstudioapi::getActiveDocumentContext()$path
output_file <- stringr::str_replace(current_file, '.Rmd', '.R')
knitr::purl(current_file, output = output_file)
file.edit(output_file)
```

### Directories
```{r, message=FALSE}
output_dir <- here('analysis/output/09_il23_profile') # analysis file output directory
data_dir <- here('data/derived/tildra') # data file output directory

dir.create(output_dir, showWarnings = FALSE)
dir.create(data_dir, showWarnings = FALSE)
```

### Loading the libraries
```{r, message=FALSE}
library(Seurat) 
library(tidyverse)
library(ggExtra)
library(SeuratWrappers)
library(wutilities) # devtools::install_github('symbiologist/wutilities')
library(tictoc)
library(ggforce)
library(ggradar) # devtools::install_github("ricardo-bion/ggradar")
library(extrafont)
library(patchwork)

theme_set(theme_dwu()) # set default theme
```
### Functions
```{r}
calculate_signature <- function(seuratobj, 
                                genes,
                                name = 'signature') {
  
  gene_data <- FetchData(seuratobj, genes)
  as.data.frame(rowSums(gene_data)) %>% set_names(name)
  
}
```

### Import data
#### scRNA-seq
```{r}
seuratobj <- read_rds(here(data_dir, 'seuratobj_annotated.rds'))
seuratobj
```
#### Visium
```{r}
visiumobj <- read_rds(here(data_dir, 'visium_integrated.rds'))
DefaultAssay(visiumobj) <- 'SCT'

visium_meta <- visiumobj@meta.data %>% rownames_to_column() %>% 
  mutate(cluster = case_when(seurat_clusters == 3 ~ 'Keratinocyte',
                             seurat_clusters == 2 ~ 'Vascular',
                             seurat_clusters == 1 ~ 'Fibroblast',
                             seurat_clusters == 0 ~ 'Stromal'))
```

### Use global or local identities and set as cluster/supercluster in the object
```{r}
which_identity <- 'local'
which_cluster <- paste0(which_identity, '_cluster')
which_supercluster <- paste0(which_identity, '_supercluster')

seuratobj$cluster <- seuratobj[[which_cluster]] # set cluster as main annotation
seuratobj$supercluster <- seuratobj[[which_supercluster]] # set cluster as main annotation
Idents(seuratobj) <- 'cluster'

output_dir <- here(output_dir, which_identity)
dir.create(output_dir)
```

### Prepare metadata
```{r}
metadata <- seuratobj@meta.data %>% rownames_to_column()

metadata_sample <- metadata %>% 
  select(id, patient, sample, condition, treatment, response, group, patient_label, sample_label) %>% 
  unique() %>% 
  as_tibble()

metadata_sample
```

#### Combine visium and scRNA metadata
```{r}
metadata_combined <- metadata %>% 
  bind_rows(visium_meta %>% select(rowname, id, cluster) %>% drop_na()) %>% 
  select(rowname, 
         id, 
         sample, 
         patient,
         condition, 
         treatment, 
         group, 
         cluster, 
         supercluster, 
         celltype, 
         sample_label, 
         patient_label) %>% 
  mutate(cellgroup = case_when(
    str_detect(cluster, 'Trm1') ~ 'Trm1',
    celltype == 'APC' ~ 'APC',
    str_detect(cluster, 'T') ~ 'Tcell',
    TRUE ~ cluster
  ))

metadata_combined
```

### Major clusters
```{r}
clusters_all <- metadata$cluster %>% unique()
```

## Gene sets
### PV50
```{r}
pv50 <- readxl::read_excel(here('analysis/input/PV50.xlsx'))
pv50_genes <- pv50$...1 %>% intersect(rownames(seuratobj))
```

### Tildra-responsive genes
#### Genes that change in responders but not non-responders
##### Group-level
```{r}
de_pre_vs_mid_response <- bind_rows(
  read_tsv(here('analysis/output/08_de', which_cluster, 'de_pre_vs_mid_responder.tsv.gz')) %>% 
    mutate(response = 'Responder'),
  read_tsv(here('analysis/output/08_de', which_cluster, 'de_pre_vs_mid_nonresponder.tsv.gz')) %>% 
    mutate(response = 'Non-responder'))

il23_group_response <- de_pre_vs_mid_response %>% 
  filter(p_val_adj < 0.05,
         avg_log2FC > 0,
         response == 'Responder') 

il23_group_response
```

##### Patient-level
```{r}
de_pre_vs_mid_patients <- read_tsv(here('analysis/output/08_de', which_cluster,'de_pre_vs_mid_per_patient.tsv.gz'))

de_patient_tally <- de_pre_vs_mid_patients %>% 
  filter(p_val_adj < 0.05,
         avg_log2FC > 0) %>% 
  group_by(cluster, gene, group) %>% 
  tally() %>% 
  pivot_wider(names_from = group,
              values_from = n,
              values_fill = 0)

de_patient_tally
```

```{r}
il23_response <- il23_group_response %>% 
  left_join(de_patient_tally)

il23_response
```

### Response genes per cluster
```{r}
il23_genes_all <- map(clusters_all %>% set_names(.), function(i) {
  
  il23_response %>% 
    filter(cluster == i, 
           p_val_adj < 0.05, 
           Responder > `Non-responder`) %>% 
    pull(gene)
  
})

il23_genes_strict <- map(clusters_all %>% set_names(.), function(i) {
  
  il23_response %>% 
    filter(cluster == i, 
           p_val_adj < 0.05, 
           Responder > `Non-responder`,
           Responder > 2, 
           `Non-responder` < 2) %>% 
    pull(gene) 
  
})

il23_genes_strict$Tcm
```

```{r}
il23_genes_table <- map(names(il23_genes_all), function(i) {
  
  strict <- il23_genes_strict[[i]]
  
  output <- tibble('cluster' = i,
         'gene' = il23_genes_all[[i]]) 
  
  if(!is.null(strict)) {
    output <- output %>% mutate(Status = ifelse(gene %in% strict, 'Strict', 'All'))
  } else {
    output <- output %>% mutate(Status = 'All')
  }
  
  output %>% mutate(across(everything(), as.character))
  
})  %>% bind_rows()

il23_genes_table %>% write_tsv(here(output_dir, 'il23_genes_table.tsv'))
il23_genes_table
```


### IL17 responsive genes from Harirchian et al. JID 2019
https://pubmed.ncbi.nlm.nih.gov/30543901/
```{r}
il17 <- bind_rows(
  readxl::read_xlsx(here('analysis/input/NIHMS1522377-supplement-2 Cho IL17A.xlsx'), sheet = 'Table S1C', skip = 1) %>% mutate(time = 'T_1'),
  readxl::read_xlsx(here('analysis/input/NIHMS1522377-supplement-2 Cho IL17A.xlsx'), sheet = 'Table S1D', skip = 1) %>% mutate(time = 'T_24'))
colnames(il17)[1] <- 'gene'
il17
```


```{r}
il17_responsive <- il17 %>% 
  filter(logFC > 0.4,
         FDR < 0.05) %>% 
  add_count(gene) %>% filter(n > 1)

il17_responsive
```


```{r}
il17_genes <- il17_responsive$gene %>% unique()

```

#### Other signatures of interest
```{r}
secreted_signature <- c('IL17A', 'IL17F', 'IL26', 'CXCl13', 'GZMB', 'GNLY', 'PCSK7')

il23_peripheral <- c('GNLY',
                     'HGPD',
                     'IL17A',
                     'IL17F',
                     'IL26',
                     'PTPN13',
                     'S100A4')
```

### Extract all expression data
```{r}
extract_expression <- function(seuratobj,
                               cells = NULL,
                               genes,
                               normalize = TRUE,
                               log = FALSE,
                               assay = 'RNA',
                               slot = 'counts') {
  
  data_subset <- FetchData(object = seuratobj, 
                           cells = cells,
                           vars = genes,
                           assay = assay,
                           slot = slot) %>% 
    rownames_to_column() %>% 
    pivot_longer(cols = -rowname,
                 names_to = 'gene',
                 values_to = 'counts')
  
  total_count <- colSums(GetAssayData(seuratobj, 
                                        slot = slot, 
                                        assay = assay))
  
  if(!is.null(cells)) {
    total_count <- total_count[cells]
  }
  
  data_subset <- data_subset %>% 
    left_join(data.frame('total' = total_count) %>% rownames_to_column(), by = 'rowname')
  
  if(normalize) {
   data_subset <- data_subset %>% 
      mutate(cpm = 1e6 * counts / total) 
  }
  
  data_subset
  
}

pseudobulk <- function(seuratobj,
                       genes,
                       cells = NULL,
                       group_by = 'ident',
                       pseudocount = 1,
                       rounding = 3,
                       assay = 'RNA',
                       slot = 'count') {
  
  metadata_groups <- FetchData(seuratobj, vars = group_by, cells = cells) %>% rownames_to_column()
  
  expression <- extract_expression(seuratobj,
                                   genes = genes,
                                   cells = cells,
                                   normalize = FALSE,
                                   log = FALSE,
                                   assay = assay,
                                   slot = slot) %>% 
    left_join(metadata_groups, by = 'rowname')
  
  totals <- expression %>% 
    select(rowname, total, all_of(group_by)) %>% 
    unique() %>% 
    group_by(across(all_of(group_by))) %>% 
    summarize(total = sum(total))
  
  merged <- expression %>% 
    group_by(across(all_of(c('gene', group_by)))) %>% 
    summarize(counts = sum(counts)) %>% 
    ungroup() %>% 
    left_join(totals) %>% 
    mutate(cpm = counts/total * 1e6,
           log10cpm = log10(cpm + pseudocount))
  
  merged %>% mutate(across(contains('cpm'), round, rounding))
  
}

plot_radar <- function(df,
                       group_by = NULL,
                       features = NULL,
                       colors = NULL,
                       title = '',
                       legend.position = 'top') {
  
  if(is.null(group_by)) {
    
    group_by <- colnames(df)[1]
    
  }
    
  if(is.null(features)) {
    
    features <- colnames(df) %>% setdiff(group_by)
    
  }
  
  radar_input <- df %>% 
    select(all_of(c(group_by, features))) %>% 
    mutate(across(all_of(features), scales::rescale_max))
  
  radar_input[[group_by]] <- droplevels(radar_input[[group_by]])
  
  radar_input %>% 
    ggradar(legend.text.size = 8,
            grid.label.size = 0,
            axis.label.size = 2,
            group.point.size = 2,
            axis.label.offset = 1.1,
            group.line.width = 1,
            group.colours = colors,
            plot.title = title,
            legend.position = legend.position) +
    theme(plot.title = element_text(size = 9, hjust = 0.5),
          plot.background = element_rect(color = 'black'),
          #legend.box.background = element_rect(color = 'blue'),
          legend.key.height = unit(1, 'mm'),
          legend.spacing = unit(0, 'mm')) 
}

plot_radar_patient <- function(pseudo,
                               cluster,
                               patient,
                               genes,
                               normalize = TRUE,
                               scale = c('within', 'across'),
                               colors = c('dodgerblue4', 'orangered'),
                               test = TRUE,
                               significance_colors = c('orangered', 'red4'),
                               significance_cutoff = 0.01,
                               title = 'patient_label') {
  
  plot_title <- pseudo %>% filter(patient == {{patient}}) %>% pull({{title}}) %>% unique()
  
  plot_input <- pseudo %>% 
    filter(cluster %in% {{cluster}}, 
           gene %in% {{genes}}) %>% 
    select(patient,
           feature = gene,
           treatment,
           normalized = cpm)
  
  # significance testing
  if(test) {
    res <- test_pre_vs_mid(pseudo = {{pseudo}},
                           cluster = {{cluster}},
                           genes = {{genes}},
                           patient = {{patient}})
    
    subtitle <- paste0('t p = ', round(res$t$p.value, 3), ', Wilcoxon p = ', round(res$wilcox$p.value, 3))
    
    if(res$t$p.value < significance_cutoff) {
      plot_title <- paste0(plot_title, '*')
    }
    
    if(res$wilcox$p.value < significance_cutoff) {
      plot_title <- paste0(plot_title, '*')
    }
    
    # color title based on significance
    n_significant <- str_count(plot_title, '\\*')
    if(n_significant == 0) {
      title_color <- 'black'
    } else {
      title_color <- significance_colors[n_significant]
    }
    
  } else {
    
    subtitle <- NULL
  
    }
  
  if(scale == 'across') {
    
    plot_input <- plot_input %>% 
      group_by(feature) %>% 
      mutate(feature_max = max(normalized),
             normalized = ifelse(feature_max > 0, normalized/feature_max, 0)) %>% 
      filter(patient == {{patient}})
    
  } else if(scale == 'within') {
    
    plot_input <- plot_input %>% 
      filter(patient == {{patient}}) %>% 
      group_by(feature) %>% 
      mutate(feature_max = max(normalized),
             normalized = ifelse(feature_max > 0, normalized/feature_max, 0))
  }
  
  radar_input <- plot_input %>% 
    mutate(treatment = droplevels(treatment)) %>% 
    select(treatment, feature, normalized) %>% 
    pivot_wider(names_from = feature,
                values_from = normalized)
  
  p_radar <- radar_input %>% 
    ggradar(legend.text.size = 8,
            grid.label.size = 0,
            axis.label.size = 2,
            group.point.size = 2,
            axis.label.offset = 1.1,
            group.line.width = 1,
            group.colours = colors,
            legend.position = 'top', 
            font.radar = 'Arial') +
    theme(text = element_text(family = 'Arial'),
          plot.title = element_text(size = 9, hjust = 0.5, color = title_color),
          plot.subtitle = element_text(size = 8, hjust = 0.5),
          plot.background = element_rect(color = 'black'),
          legend.key.height = unit(1, 'mm'),
          legend.spacing = unit(0, 'mm')) +
    labs(title = plot_title,
         subtitle = subtitle)
  
  p_radar
  
  
}

test_pre_vs_mid <- function(pseudo,
                            patient,
                            cluster,
                            genes,
                            logtransform = TRUE,
                            pseudocount = 1) {
  
  test_input <- pseudo %>% 
    filter(cluster %in% {{cluster}}, 
           gene %in% {{genes}},
           patient == {{patient}}) %>% 
    select(gene,
           treatment,
           expression = cpm) %>% 
    rowwise() %>% 
    mutate(expression = ifelse(logtransform, log10(expression + 1), expression)) %>% 
    ungroup() %>% 
    pivot_wider(names_from = 'treatment',
                values_from = 'expression')
  
  if(ncol(test_input) == 3) {
    t_res <- t.test(test_input$Pre,
                    test_input$Mid,
                    paired = TRUE)
    
    wilcox_res <- wilcox.test(test_input$Pre,
                              test_input$Mid,
                              paired = TRUE)
  } else {
    t_res <- list(p.value = 1)
    wilcox_res <- list(p.value = 1)
  }
  
  list('t' = t_res,
       'wilcox' = wilcox_res)
}


```

```{r}
tildra_barcodes <- metadata %>% filter(treatment %in% c('Pre', 'Mid')) %>% pull(rowname)
genes_core <- c('IL17A', 'IL17F', 'IL26', 'IFNG')
gene_set <- c('IL17A', 'IL17F', 'IL26', 'TNF', 'IFNG', 'IL23R', 'IL21', 'IL22', 'IL36A', 'CXCL13')

pseudo <- pseudobulk(seuratobj,
                     cells = tildra_barcodes,
                     genes = c(genes_core, gene_set, il23_genes_all %>% unlist()) %>% unique(),
                     group_by = c('cluster', 'id')) %>% 
  left_join(metadata_sample %>% select(id, patient, treatment, group, sample_label, patient_label))

pseudo 
```

### Radar plot

```{r}

p <- plot_radar_patient(pseudo = pseudo,
                   cluster = 'Mac1',
                   patient = 'P1',
                   genes = c(genes_core, il23_genes_strict$Mac1),
                   scale = 'across')

# ggsave(plot = p,
#        h = 10,
#        w = 10,
#        device = cairo_pdf,
#        path = output_dir,
#        filename = 'test.pdf')

p
```

#### Limited set
```{r}
p <- map(pseudo$patient %>% unique() %>% sort() %>% as.character(), function(which_patient) {
  plot_radar_patient(pseudo = pseudo,
                     cluster = 'Trm1',
                     patient = which_patient,
                     genes = gene_set,
                     scale = 'within')
}) %>% wrap_plots(ncol = 3)

ggsave(plot = p,
       h = 10,
       w = 10,
       device = cairo_pdf,
       path = output_dir,
       filename = 'radar_cytokines.pdf')

p
```


#### Loop
```{r}
update_geom_defaults('text', list(fontface = 'italic'))

map(pseudo$cluster %>% unique() %>% sort() %>% as.character(), function(which_cluster) {
  
  genes_all <- c(genes_core, il23_genes_all[[which_cluster]]) %>% unique()
  genes_strict <- c(genes_core, il23_genes_strict[[which_cluster]]) %>% unique()
  
  map(c('across', 'within'), function(which_scale) {
    
    if(length(genes_all) > length(genes_core)) {
      
      print2(paste(which_cluster, which_scale))
      
      p <- map(pseudo$patient %>% unique() %>% sort() %>% as.character(), function(which_patient) {
        
        
        if((pseudo %>% 
            filter(patient == which_patient,
                   cluster == which_cluster,
                   gene %in% genes_all,
                   cpm > 0) %>% 
            nrow() > 0)) {
          
          plot_radar_patient(pseudo = pseudo,
                             cluster = which_cluster,
                             patient = which_patient,
                             genes = genes_all,
                             scale = which_scale)
        } else {
          ggplot()
        }
      }) %>% wrap_plots() 
      
      print2('Saving')
      dir.create(here(output_dir, 'radar_all'), showWarnings = FALSE)
      ggsave(plot = p,
             h = 10,
             w = 10,
             device = cairo_pdf,
             path = here(output_dir, 'radar_all'),
             filename = paste0(which_scale, '_', str_remove(which_cluster, '/'), '.pdf'))
      
    }
    
    if(length(genes_strict) > length(genes_core)) {
      
      p <- map(pseudo$patient %>% unique() %>% sort(), function(which_patient) {
        
        
        if((pseudo %>% filter(patient == which_patient,
                              cluster == which_cluster,
                              gene %in% genes_strict,
                              cpm > 0) %>% 
            nrow() > 0)) {
          
          plot_radar_patient(pseudo = pseudo,
                             cluster = which_cluster,
                             patient = which_patient,
                             genes = genes_strict,
                             scale = which_scale)
        } else {
          ggplot()
        }
        
        
      }) %>% wrap_plots()
      
      dir.create(here(output_dir, 'radar_strict'), showWarnings = FALSE)
      
      ggsave(plot = p,
             h = 10,
             w = 10,
             device = cairo_pdf,
             path = here(output_dir, 'radar_strict'),
             filename = paste0(which_scale, '_', str_remove(which_cluster, '/'), '.pdf'))
    }
  }) %>% invisible()
}) %>% invisible()

update_geom_defaults('text', list(fontface = 1))

```

### Stats
```{r}
which_cluster <- 'Trm1'
genes_all <- c(genes_core, il23_genes_all[[which_cluster]]) %>% unique()
genes_strict <- c(genes_core, il23_genes_strict[[which_cluster]]) %>% unique()
  
stats_input <- pseudo %>% 
  filter(cluster == which_cluster,
         gene %in% genes_all) %>% 
  select(gene, patient, treatment, group, cpm) %>% 
  pivot_wider(names_from = treatment,
              values_from = cpm) %>% 
  mutate(ratio = (Pre - Mid)/Pre) %>% 
  arrange(gene, patient) %>% 
  group_by(gene, group) %>% 
  summarize(ratio = median(ratio)) %>% 
  pivot_wider(names_from = group,
              values_from = ratio)

stats_input
```
```{r}
stats_input %>% 
  ungroup() %>% 
  summarize(across(where(is.numeric), median))
```

```{r}
wilcox.test(stats_input$Responder,
            stats_input$`Non-responder`,
            paired = TRUE)
```



```{r}
extract_signature <- function(seuratobj,
                              cells = NULL,
                              genes,
                              name = 'signature',
                              normalize = TRUE,
                              log = FALSE,
                              assay = 'RNA',
                              slot = 'counts') {
  
  seuratobj <- seuratobj %>% add_signature(genes = genes, name = name)
  
}

add_signature <- function(seuratobj, 
                          genes,
                          name = 'signature') {
  
  seuratobj <- seuratobj %>% PercentageFeatureSet(features = genes, col.name = name)
  seuratobj[[name]] <- seuratobj[[name]] * seuratobj[['nCount_RNA']]/100
  
  seuratobj
}
```


### IL23-responsive gene expression in mid-treatment samples 
Strict
```{r}
p <- data_table_long_tx_summarized %>% 
  filter(gene %in% c(il23_genes_strict, 'IL17A', 'IL17F'),
         treatment == 'Mid',
         cellgroup == 'Trm1') %>% 
  ggplot(aes(x = reorder(gene, expression),
             y = expression,
             color = group)) +
  #geom_boxplot(outlier.color = NA) +
  geom_bar(stat = 'summary', position = 'dodge', fill = 'grey90') +
  #facet_wrap(~gene, scales = 'free_x') +
  theme_dwu(legend.position = 'top') +
  scale_color_manual(values = c('orangered', 'dodgerblue4')) +
  geom_jitter(alpha = 0.8, size = 2) +
  coord_flip() +
  labs(x = '',
       y = 'Trm1 Average Expression')

ggsave(plot = p,
       filename = 'bar_mid_separators_strict.pdf',
       path = output_dir,
       h = 14,
       w = 6)

p

```
Full
```{r}
p <- data_table_long_tx_summarized %>% 
  
  filter(gene %in% c(il23_genes_all, 'IL17A', 'IL17F'),
         treatment == 'Mid',
         cellgroup == 'Trm1') %>% 
  ggplot(aes(x = reorder(gene, expression),
             y = expression,
             color = group)) +
  #geom_boxplot(outlier.color = NA) +
  geom_bar(stat = 'summary', position = 'dodge', fill = 'grey90') +
  geom_sina(alpha = 0.8, size = 2) +
  #facet_wrap(~gene, scales = 'free_x') +
  theme_dwu(legend.position = 'top') +
  scale_color_manual(values = c('orangered', 'dodgerblue4')) +
  coord_flip() +
  labs(x = '',
       y = 'Trm1 Average Expression')

ggsave(plot = p,
       filename = 'bar_mid_separators_full.pdf',
       path = output_dir,
       h = 44,
       w = 6)

p
```
### Facets
```{r}
p <- data_table_long_tx_summarized %>% 
  
  filter(gene %in% c(il23_genes_all, 'IL17A', 'IL17F'),
         treatment == 'Mid',
         cellgroup == 'Trm1') %>% 
  ggplot(aes(x = group,
             y = expression,
             color = group)) +
  #geom_boxplot(outlier.color = NA) +
  geom_bar(stat = 'summary', position = 'dodge', fill = 'grey90') +
  geom_point(alpha = 0.8, size = 2, aes(color = patient)) +
  facet_wrap(~gene, scales = 'free_y', ncol = 15) +
  theme_dwu(legend.position = 'top') +
  rcartocolor::scale_color_carto_d() +
  #scale_color_manual(values = c('orangered', 'dodgerblue4')) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) + 
  labs(x = '',
       y = 'Trm1 Average Expression')

ggsave(plot = p,
       filename = 'bar_mid_separators_facet_full.pdf',
       path = output_dir,
       h = 30,
       w = 17)
```

```{r}
p <- data_table_long_tx_summarized %>% 
  
  filter(gene %in% c(il23_genes_strict, 'IL17A', 'IL17F'),
         treatment == 'Mid',
         cellgroup == 'Trm1') %>% 
  ggplot(aes(x = group,
             y = expression,
             color = group)) +
  #geom_boxplot(outlier.color = NA) +
  geom_bar(stat = 'summary', position = 'dodge', fill = 'grey90') +
  geom_point(alpha = 0.8, size = 2, aes(color = patient)) +
  facet_wrap(~gene, scales = 'free_y', ncol = 10) +
  theme_dwu(legend.position = 'top') +
  #scale_color_manual(values = c('orangered', 'dodgerblue4')) +
  rcartocolor::scale_color_carto_d(palette = 'Prism') +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) + 
  labs(x = '',
       y = 'Trm1 Average Expression')

ggsave(plot = p,
       filename = 'bar_mid_separators_facet_strict.pdf',
       path = output_dir,
       h = 14,
       w = 12)
```




#### Secretion
```{r}
p <- data_table_long_tx_summarized %>% 
  
  filter(gene %in% 'Secreted',
         treatment == 'Mid',
         cellgroup == 'Trm1') %>% 
  ggplot(aes(x = reorder(patient, -expression),
             y = expression,
             fill = group)) +
  #geom_boxplot(outlier.color = NA) +
  geom_bar(stat = 'summary', position = 'dodge', color = 'black') +
  theme_dwu(legend.position = 'top') +
  #rcartocolor::scale_color_carto_d(palette = 'Prism') +
  scale_fill_manual(values = c('orangered', 'dodgerblue4')) +
  labs(x = 'Patient',
       y = 'Trm1 Average Expression',
       title = 'Secreted Gene Signature')

p
```

#### Peripheral Th17 IL23
```{r}


p <- data_table_long_tx_summarized %>% 
  
  filter(gene == 'IL23_peripheral',
         treatment == 'Mid',
         cellgroup == 'Trm1') %>% 
  ggplot(aes(x = reorder(patient, -expression),
             y = expression,
             fill = group)) +
  #geom_boxplot(outlier.color = NA) +
  geom_bar(stat = 'summary', position = 'dodge', color = 'black', fun = sum) +
  theme_dwu(legend.position = 'top') +
  #rcartocolor::scale_color_carto_d(palette = 'Prism') +
  scale_fill_manual(values = c('orangered', 'dodgerblue4')) +
  labs(x = 'Patient',
       y = 'Trm1 Average Expression',
       title = 'Peripheral IL23 Gene Signature')

p
```

### Statistics
```{r}

```
### T vs APC
```{r}
p <- data_table_long %>% 
  filter(treatment != 'None',
         gene %in% c('IL17A', 'IL17F'),
         group %in% c('APC', 'Trm1', 'Tcell')) %>% 
  group_by(patient, treatment, group, group, gene) %>% 
  summarize(expression = mean(expression)) %>% 
  ggplot(aes(x = patient,
             y = expression,
             fill = group)) +
  geom_bar(stat = 'identity') +
  facet_wrap(~group + treatment + gene) +
  theme_dwu(legend.position = 'top')
  
p
```


### PV, AD, HC
```{r fig.height=6, fig.width=12}
display_features <- c('APC IL23A',
                      'Trm1 IL23R',
                      paste('Trm1', il23_genes_all),
                      'Trm1 IL17A',
                      'Trm1 IL17F',
                      'APC IL17A',
                      'APC IL17F',
                      'Trm1 ZFP36',
                      'Trm1 ZFP36L2') %>% 
  unique()

plot_input <- data_table_long_tx_summarized %>% 
  mutate(feature = paste(cellgroup, group)) %>% 
  filter(feature %in% display_features) %>% 
  mutate(feature = factor(feature, levels = display_features))

p <- plot_input %>% 
  ggplot(aes(x = feature,
             y = normalized,
             fill = condition)) +
  geom_boxplot()+
  theme_dwu(legend.position = 'top') +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = 'Normalized Expression',
       x = '') +
  ggthemes::scale_fill_few()

p
```

```{r}

metadata_combined %>% 
  filter(cluster %in% clusters_apc,
         treatment != 'None') %>% 
  group_by(group, cluster, treatment) %>% 
  tally() %>% 
  pivot_wider(names_from = c(group, treatment),
              values_from = n) %>% 
  View()

```

## Session Info
```{r}
sessionInfo()
```
