==> analysis/scripts/11_visium_analysis.Rmd <==
---
title: "Visium Analysis"
author: "David Wu"
---


### Use the here package for robust relative paths
```{r}
library(here)
```

### Optionally convert this Rmd to R script 
```{r eval=FALSE, message=FALSE}
current_file <- rstudioapi::getActiveDocumentContext()$path
output_file <- stringr::str_replace(current_file, '.Rmd', '.R')
knitr::purl(current_file, output = output_file)
file.edit(output_file)
```

### Directories
```{r, message=FALSE}
output_dir <- here('analysis/output/11_visium_analysis') # analysis file output directory
data_dir <- here('data/derived/tildra') # data file output directory

dir.create(output_dir, showWarnings = FALSE)
dir.create(data_dir, showWarnings = FALSE)
```

### Loading the libraries
```{r, message=FALSE}
library(Seurat) 
library(SeuratWrappers)

library(tidyverse)
library(ggthemes)
library(patchwork)
library(ggforce)
library(ggExtra)

library(tictoc)
library(wutilities) # devtools::install_github('symbiologist/wutilities')
theme_set(theme_dwu())
source(here('figures/colors.R')) # consistent color palettes
```

### Import Visium object and add metadata
```{r}
seuratobj <- read_rds(here(data_dir, 'visium_integrated.rds'))
seuratobj
```

```{r}
DefaultAssay(seuratobj)
```

### Organize patient metadata
```{r}
patient_metadata <- seuratobj@meta.data %>% 
  select(id, sample, patient, treatment, response, group, ends_with('label')) %>% unique() %>% as_tibble() %>% 
  mutate(treatment = factor(treatment, levels = c('Pre', 'Mid'))) %>% 
  mutate(category = case_when(
    patient == 'P7' ~ 'Non-responder (Trm-sensitive)',
    group == 'Responder' ~ 'Responder',
    group == 'Non-responder' ~ 'Non-responder (Trm-insensitive)')) %>% 
  mutate(category = factor(category, 
                           levels = c('Responder', 'Non-responder (Trm-sensitive)', 'Non-responder (Trm-insensitive)'))) %>% 
  arrange(patient, treatment) 

seuratobj$sample <- factor(seuratobj$sample, 
                           levels = patient_metadata$sample)
```

```{r fig.height = 12, fig.width = 12}
p <- SpatialDimPlot(seuratobj, 
                    ncol = 4, 
                    image.alpha = 0) + theme(legend.position = 'top') 
p
```

### Markers
```{r}
seuratobj <- seuratobj %>% PrepSCTFindMarkers()
markers <- RunPrestoAll(seuratobj, assay = 'SCT') %>% mutate(pct.diff = pct.1 - pct.2)
```
```{r}
markers %>% 
  filter(cluster == 0) %>% 
  arrange(p_val)
```

```{r}
markers %>% 
  filter(cluster == 1) %>% 
  arrange(-pct.diff)
```
```{r}
markers %>% 
  filter(cluster == 2) %>% 
  arrange(-pct.diff)
```

```{r}
markers %>% 
  filter(cluster == 3) %>% 
  arrange(-pct.diff)
```
```{r}
seurat_feature(seuratobj)
```
```{r}
seurat_feature(seuratobj, features = c('KRT1','KRT2', 'KRT10'))
```
```{r}
VlnPlot(seuratobj, features = c('KRT1', 'KRT2', 'KRT10'))
```
```{r}
VlnPlot(seuratobj, features = c('MCAM', 'PECAM1', 'TAGLN'))
```
```{r}
VlnPlot(seuratobj, features = c('COL1A1', 'COL6A1'))
```

```{r fig.height = 12, fig.width = 16}
p <- SpatialFeaturePlot(seuratobj,
                        features = 'KRT10',
                        image.alpha = 0,
                        ncol = 4)
p
```


```{r fig.height = 12, fig.width = 16}
p <- SpatialFeaturePlot(seuratobj,
                        features = 'MCAM',
                        ncol = 4)

p
```
```{r fig.height = 12, fig.width = 16}
p <- SpatialFeaturePlot(seuratobj,
                        features = 'COL1A1',
                        ncol = 4, image.alpha = 0)

p
```
```{r fig.height=12, fig.width=14}
p <- SpatialFeaturePlot(seuratobj,
                        features = 'CD3D',
                        ncol = 4)

p
```

```{r fig.height=12, fig.width=14}
p <- SpatialFeaturePlot(seuratobj,
                        features = 'PTPRC',
                        ncol = 4, 
                        slot = 'counts')

p
```


## Counts analysis
```{r}
gene_set <- c('IL13', 'IL17A', 'IL17F', 'IL26', 'IL21', 'IL22', 'IL23A', 'IL23R', 'IL36A', 'CXCL13', 'TNF', 'IFNG')
genes_of_interest <- c('CD3D', 'CD4', 'CD8A', 'CD8B', 'PTPRC')
genes_control <- c('ACTB', 'COL1A1', 'KRT10', 'COL1A2', 'S100A9', 'UBC', 'FABP5', 'B2M', 'SPARC', 'COL6A1', 'COL6A2', 'DCN', 'MYH9')
gene_counts <- rowSums(seuratobj@assays$Spatial@counts) %>% sort(decreasing = TRUE)

genes_keratinocyte <- c('DEFB4A',
                        'CXCL1',
                        'PI3',
                        'S100A7A',
                        'SPRR2B',
                        'SERPINB3',
                        'IL36G',
                        'SERPINB4',
                        'RHCG',
                        'SLC6A14',
                        'PLAT',
                        'ZC3H12A',
                        'LCN2',
                        'IL19',
                        'DEFB103B',
                        'SPRR2E',
                        'CCL3',
                        'SPRR2A',
                        'PRSS22',
                        'SPRR2F',
                        'SPRR2D')

genes_stromal <- c('CCL8',
                   'CXCL1',
                   'CXCL2',
                   'CXCL3',
                   'CXCL8',
                   'CCL20',
                   'CXCL6',
                   'CCL7',
                   'CCL2',
                   'AREG',
                   'NAMPT',
                   'BMP2',
                   'NDP',
                   'CSF2',
                   'IL6')

#biocarta Th1 TH2 genes
genes_th1th2 <- c("IL4R",
                  "CD40LG",
                  "IFNGR1",
                  "IL2RA",
                  "IL4R",
                  "IL2",
                  "IL4",
                  "IFNG",
                  "IL12A",
                  "CD86",
                  "CD28",
                  "IL18",
                  "HLA-DRB1",
                  "CD40",
                  "IL4R",
                  "IL12RB2",
                  "IL18",
                  "HLA-DRB1",
                  "HLA-DRB5",
                  "IL12B",
                  "IL18R1",
                  "NR1I2",
                  "IFNGR2",
                  "IL12RB1",
                  "CD28",
                  "CD86",
                  "HLA-DRA",
                  "HLA-DRB4",
                  "NR1I2",
                  "HLA-DRB3",
                  "NR1I2",
                  "CD40",
                  "IL12RB1",
                  "IL4",
                  "CD86")

# ray th1 th2 genes
genes_th1 <- c('CXCL9', 'CXCL10', 'CXCL11', 'ICAM1', 'TNF', 'CXCL8', 'CCL5')
genes_th2 <- c('CXCL1', 'OASL', 'KRT16', 'CLDN8', 'TSLP', 'IL25', 'IL33', 'IL1RL1')
genes_th1th2 <- c(genes_th1, genes_th2)

genes_il17 <- rownames(seuratobj) %>% str_subset('IL17')
```

### Annotate
Keratinocytes = cluster 3
Fibroblasts = cluster 0

```{r}
counts_table <- FetchData(seuratobj,
                          vars = c('id', 'sample', 'ident', 'nCount_Spatial', paste0('spatial_', c(gene_set, genes_of_interest, genes_keratinocyte, genes_il17, genes_th1th2, genes_stromal)) %>% unique), 
                          slot = 'counts') %>% 
  rownames_to_column() %>% 
  as_tibble() %>% 
  mutate(cluster = case_when(
    spatial_CD3D + spatial_CD4 + spatial_CD8A + spatial_CD8B > 1 ~ 'Lymphoid',
    spatial_PTPRC > 2 ~ 'Myeloid',
    ident == 3 ~ 'Keratinocyte',
    ident == 2 ~ 'Vascular',
    ident == 1 ~ 'Fibroblast',
    ident == 0 ~ 'Stromal'
  )) #%>% mutate(cluster == ifelse(cluster == 'Lymphoid' & (spatial_IL17A + spatial_IL17F)  > 0, 'T17', 'Lymphoid'))


counts_table
```
```{r}
counts_table %>% 
  ggplot(aes(x = spatial_PTPRC,
             y = spatial_CD3D)) +
  geom_jitter(alpha = 0.5) +
  scale_y_continuous(limits = c(0,10),
                     breaks = c(seq(0,10))) +
  coord_equal()
```
```{r}
counts_table %>% 
  ggplot(aes(x = spatial_PTPRC)) +
  geom_histogram(bins = 20)
```


```{r}
counts_table %>% 
  group_by(sample,
           id,
           cluster) %>% 
  tally() %>% 
  pivot_wider(names_from = 'cluster',
              values_from = 'n',
              values_fill = 0) %>% 
  rowwise() %>% 
  mutate(total = sum(across(where(is.numeric))))
```


```{r}
counts_table %>% 
  group_by(id,
           cluster) %>% 
  tally() %>% 
  pivot_wider(names_from = 'cluster',
              values_from = 'n') %>% 
  rowwise() %>% 
  mutate(total = sum(across(where(is.numeric))))
```

```{r}
spot_count <- counts_table %>% 
  mutate(across(where(is.numeric), function(i) { as.integer(i > 0)})) %>% 
  add_count(id, name = 'total_spots') %>% 
  pivot_longer(cols = starts_with('spatial'),
               names_to = 'gene',
               values_to = 'positive_spots') %>% 
  group_by(id, gene, cluster) %>% 
  summarize(positive_spots = sum(positive_spots),
            total_spots = unique(total_spots)) %>% 
  mutate(gene = str_remove(gene, 'spatial_'),
         pct_spots = round(100 * positive_spots/total_spots, 2))

spot_count
  
```


```{r}
total_counts <- counts_table %>% 
  group_by(id, cluster) %>% 
  summarize(across(where(is.numeric), sum))

total_counts
```

```{r}
total_long <- total_counts %>% 
  pivot_longer(cols = starts_with('spatial'),
               names_to = 'gene',
               values_to = 'counts') %>% 
  mutate(gene = str_remove(gene, 'spatial_')) %>% 
  left_join(spot_count) %>% 
  left_join(seuratobj@meta.data %>% 
              select(id, sample, patient, treatment, response, group) %>% unique()) %>% 
  mutate(treatment_group = factor(paste(group, treatment),
                                  levels = c('Responder Pre',
                                             'Responder Mid',
                                             'Non-responder Pre',
                                             'Non-responder Mid')),
         group = factor(group, levels = c('Responder', 'Non-responder')))

total_long
```
#### Analyze counts
```{r}
total_long %>% 
  filter(gene %in% gene_set) %>% 
  ggplot(aes(x = treatment_group,
             y = counts,
             fill = group)) +
  geom_boxplot() +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  scale_x_discrete(labels = c('Pre', 'Mid', 'Pre', 'Mid')) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = 'Total UMI / biopsy',
       x = '')
```
```{r}
total_long %>% 
  filter(gene %in% gene_set) %>% 
  ggplot(aes(x = treatment_group,
             y = positive_spots,
             fill = group)) +
  geom_boxplot() +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  scale_x_discrete(labels = c('Pre', 'Mid', 'Pre', 'Mid')) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = 'Positive Spots / biopsy',
       x = '')
```
```{r}
total_long %>% 
  filter(gene %in% gene_set) %>% 
  ggplot(aes(x = treatment_group,
             y = pct_spots,
             fill = group)) +
  geom_boxplot() +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  scale_x_discrete(labels = c('Pre', 'Mid', 'Pre', 'Mid')) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = '% Positive Spots / biopsy',
       x = '')
```

### Patient level
```{r}
p <- total_long %>% 
  filter((gene %in% gene_set)) %>%
  filter(cluster %in% c('Keratinocyte')) %>% 
  ggplot(aes(x = sample,
             y = counts,
             fill = group)) +
  # ggplot(aes(x = sample,
  #            y = counts,
  #            fill = cluster)) +
  geom_col(color = 'black') +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  #scale_x_discrete(labels = c('Pre', 'Mid', 'Pre', 'Mid')) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = 'Total UMI / biopsy',
       x = '')
p
```
### Keratinocyte IL17 response
https://www.nature.com/articles/s41467-022-35319-w#MOESM1
```{r}
p <- total_long %>% 
  filter(gene %in% setdiff(genes_keratinocyte, 'CXCL1')) %>% 
    filter(cluster %in% c('Keratinocyte')) %>% 
  ggplot(aes(x = sample,
             y = counts,
             fill = group)) +
  geom_col(color = 'black') +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = 'Total UMI / biopsy',
       x = '')
p
```
```{r}

```

#### All IL17 genes
```{r}
p <- total_long %>% 
  filter(gene %in% genes_il17) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  ggplot(aes(x = sample,
             y = counts,
             fill = group)) +
  
  geom_col(color = 'black') +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = 'Total UMI / biopsy',
       x = '')
p
```
```{r}
p <- total_long %>% 
  filter(gene %in% genes_il17) %>% 
  ggplot(aes(x = sample,
             y = pct_spots,
             fill = group)) +
  geom_col(color = 'black') +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = '% Positive Spots',
       x = '')
p
```
```{r}
counts_table %>% 
  ggplot(aes(x = spatial_CD3D,
             y = spatial_IL17A)) +
  geom_jitter(alpha = 0.5)
```


#### Spatial coordinates
```{r}
spatial_neighbors <- function(seuratobj,
                              barcode,
                              distance = 1,
                              neighbors_only = FALSE) {
  
  data_spatial <- seuratobj@meta.data %>% 
    rownames_to_column() %>% 
    select(rowname, id, sample, patient, treatment, group, row, col, imagerow, imagecol) %>% 
    as_tibble() 
  
  data_barcode <- data_spatial %>% 
    filter(rowname == barcode)
  
   data_neighbors <- data_spatial %>% 
    filter(id == data_barcode$id) %>%
     mutate(row_dist = imagerow - data_barcode$imagerow,
            col_dist = imagecol - data_barcode$imagecol,
            total_dist = sqrt(row_dist^2 + col_dist^2),
            distance = ceiling(total_dist/327)) %>% 
     filter(distance <= {{distance}}) %>% 
     arrange(total_dist)
  
  if(neighbors_only) {
    data_neighbors %>% filter(barcode != {{barcode}})
  } else {
    data_neighbors  
  }
}
```

## Test neighbors
```{r}
test_spatial <- FetchData(seuratobj, vars = c('row', 'col', 'imagerow','imagecol','patient', 'treatment', 'id', 'spatial_CD3D'), slot = 'counts') %>% rownames_to_column()
test_spatial %>% arrange(-spatial_CD3D) %>% filter(id == 'S292')
```


```{r}
barcode <- 'CAGTGTCGGCTGGCCC-1_10'
neighbors <- spatial_neighbors(seuratobj,
                               barcode = barcode,
                               distance = 20)

neighbors
```
```{r}
poi <- neighbors %>% filter(rowname == barcode)
plot_input <- neighbors
plot_input %>% 
  select(-c(id:group)) %>% 
  mutate(total_dist = round(total_dist))
```

```{r}
dists <- plot_input %>% 
  mutate(total_dist = round(total_dist)) %>% 
  group_by(total_dist, distance) %>% 
  tally()
  
dists 
```

```{r}
plot_input %>% 
  ggplot(aes( x = imagecol, y = imagerow,
              fill = factor(distance, levels = 1:20))) + 
  geom_point(shape = 21,
             size = 4) + 
  ggthemes::scale_fill_tableau(palette = 'Tableau 20') +
  coord_equal() +
  theme(legend.position = 'right')
```
```{r}
plot_input %>% 
  group_by(distance) %>% 
  summarize(min = min(total_dist),
            med = median(total_dist),
            max = max(total_dist)) %>% 
  mutate(across(where(is.numeric), round))
```

### Analyze IL17 secreting sectors
```{r fig.height = 10, fig.width=10}
p <- SpatialFeaturePlot(seuratobj,
                        features = 'CD3D',
                        ncol = 4, 
                        image.alpha = 0.1,
                        slot = 'counts')

p
```

### Lymphoid spots
```{r}
distance <- 1

lymphoid_spots <- counts_table %>% filter(cluster == 'Lymphoid') %>% pull(rowname)
lymphoid_adjacent_spots <- lymphoid_spots %>% map(function(i) {
  spatial_neighbors(seuratobj = seuratobj,
                    distance = {{distance}},
                    barcode = i) 
}) %>% bind_rows()

lymphoid_adjacent_spots
```
```{r}
counts_adjacent <- counts_table %>% 
  left_join(lymphoid_adjacent_spots %>% 
              select(rowname) %>% 
              mutate(status = 'adjacent')) %>% 
  filter(status == 'adjacent') 

counts_adjacent
```


```{r}
spot_count_adjacent <- counts_adjacent %>% 
  mutate(across(where(is.numeric), function(i) { as.integer(i > 0)})) %>% 
  add_count(id, name = 'total_spots') %>% 
  pivot_longer(cols = starts_with('spatial'),
               names_to = 'gene',
               values_to = 'positive_spots') %>% 
  group_by(id, gene, cluster) %>% 
  summarize(positive_spots = sum(positive_spots),
            total_spots = unique(total_spots)) %>% 
  mutate(gene = str_remove(gene, 'spatial_'),
         pct_spots = round(100 * positive_spots/total_spots, 2))

spot_count_adjacent
  
```


```{r}
total_counts_adjacent <- counts_adjacent %>% 
  group_by(id, cluster) %>% 
  summarize(across(where(is.numeric), sum))

total_counts_adjacent
```

```{r}
total_long_adjacent <- total_counts_adjacent %>% 
  pivot_longer(cols = starts_with('spatial'),
               names_to = 'gene',
               values_to = 'counts') %>% 
  mutate(gene = str_remove(gene, 'spatial_')) %>% 
  left_join(spot_count_adjacent) %>% 
  left_join(seuratobj@meta.data %>% 
              select(id, sample, patient, treatment, response, group) %>% unique()) %>% 
  mutate(treatment_group = factor(paste(group, treatment),
                                  levels = c('Responder Pre',
                                             'Responder Mid',
                                             'Non-responder Pre',
                                             'Non-responder Mid')),
         group = factor(group, levels = c('Responder', 'Non-responder')))

total_long_adjacent
```
```{r}
p <- total_long_adjacent %>% 
  filter(gene %in% setdiff(genes_keratinocyte, 'CXCL1')) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  ggplot(aes(x = sample,
             y = counts,
             fill = group)) +
  geom_bar(stat = 'summary',
           color = 'black',
           fun = 'sum') +
 # geom_col() +
  facet_wrap(~gene, scales = 'free_y', nrow = 4) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_x_discrete(drop = FALSE) +
  labs(y = 'Total UMI / biopsy',
       x = '')

ggsave(plot = p,
       path = output_dir,
       device = cairo_pdf,
       h = 10,
       w = 10,
       filename = paste0('keratinocyte_genes_adjacent', distance, '.pdf'))
p
```

### Pattern
```{r}
library(ggpattern)
patient_labels <- patient_metadata %>% 
  filter(!is.na(group)) %>% 
  mutate(label = ifelse(treatment == 'Pre', as.character(patient), '')) %>% 
  pull(label)

p <- total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category = group)) %>% 
  filter(gene %in% setdiff(genes_keratinocyte, 'CXCL1')) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  ggplot(aes(x = sample,
             y = counts,
             fill = category,
             pattern = treatment)) +
  geom_bar_pattern(stat = 'identity', 
                   color = 'black', 
                   alpha = 0.8,
                   pattern_density = 0.1,
                   pattern_angle = 45,
                   pattern_key_scale_factor = 0.6) +
  facet_wrap(~gene, scales = 'free_y', nrow = 4) +
  scale_pattern_manual(values = c(Pre = 'stripe', Mid = 'none')) +
  scale_x_discrete(labels = patient_labels,
                   drop = FALSE) +
  guides(pattern = guide_legend(override.aes = list(fill = 'white')),
         fill = guide_legend(override.aes = list(pattern = 'none'))) +
  scale_fill_manual(values = rev(colors_response2)) +
  theme(axis.ticks.x = element_blank(),
        axis.text.x = element_text(hjust = 0),
        legend.position = 'top') +
  labs(y = 'Total UMI / biopsy',
       x = '')

ggsave(plot = p,
       path = output_dir,
       device = cairo_pdf,
       h = 6,
       w = 10,
       filename = paste0('keratinocyte_genes_adjacent_pattern', distance, '.pdf'))
p
```
```{r}
p <- total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category = group)) %>% 
  filter(gene %in% setdiff(genes_keratinocyte, 'CXCL1')) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  group_by(group, treatment, gene) %>% 
  mutate(label = factor(paste(group, treatment),
                        levels = c(
                          'Responder Pre',
                          'Responder Mid',
                          'Non-responder Pre',
                          'Non-responder Mid')
                        )) %>% 
  ggplot(aes(x = label,
             y = counts,
             fill = group)) +
  geom_boxplot() +
  facet_wrap(~gene, nrow = 4, scales = 'free_y') +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_manual(values = rev(colors_response2)) +
  labs(x = NULL,
       y = 'Total UMI / biopsy')

ggsave(plot = p,
       path = output_dir,
       device = cairo_pdf,
       h = 6,
       w = 10,
       filename = paste0('keratinocyte_genes_adjacent_avg', distance, '.pdf'))
p
```
```{r}
p <- total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category = group)) %>% 
  filter(gene %in% genes_stromal) %>% 
  filter(cluster %in% c('Fibroblast', 'Stromal')) %>% 
  group_by(group, treatment, gene) %>% 
  mutate(label = factor(paste(group, treatment),
                        levels = c(
                          'Responder Pre',
                          'Responder Mid',
                          'Non-responder Pre',
                          'Non-responder Mid')
                        )) %>% 
  ggplot(aes(x = label,
             y = counts,
             fill = group)) +
  geom_boxplot() +
  facet_wrap(~gene, nrow = 4, scales = 'free_y') +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_manual(values = rev(colors_response2)) +
  labs(x = NULL,
       y = 'Total UMI / biopsy')

ggsave(plot = p,
       path = output_dir,
       device = cairo_pdf,
       h = 6,
       w = 10,
       filename = paste0('stromal_genes_adjacent_avg', distance, '.pdf'))
p
```


```{r}

```

### Mid only
```{r}
patient_labels <- patient_metadata %>% 
  filter(!is.na(group)) %>% 
  mutate(label = ifelse(treatment == 'Pre', as.character(patient), '')) %>% 
  pull(label)

p <- total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category, group)) %>% 
  filter(gene %in% setdiff(genes_keratinocyte, 'CXCL1')) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  filter(treatment == 'Mid') %>% 
  ggplot(aes(x = sample,
             y = counts,
             fill = group,
             pattern = treatment)) +
  geom_bar(stat = 'identity', 
                   color = 'black', 
                   alpha = 0.8) +
  facet_wrap(~gene, scales = 'free_y', nrow = 4) +
  scale_x_discrete(labels = patient_labels,
                   drop = FALSE) +
  scale_fill_manual(values = rev(colors_response2)) +
  theme(axis.ticks.x = element_blank(),
        axis.text.x = element_text(hjust = 0),
        legend.position = 'top') +
  labs(y = 'Total UMI / biopsy',
       x = '')

ggsave(plot = p,
       path = output_dir,
       device = cairo_pdf,
       h = 6,
       w = 10,
       filename = paste0('keratinocyte_genes_adjacent_mid', distance, '.pdf'))
p
```

Check specific genes
```{r}

patient_labels <- patient_metadata %>% 
  filter(!is.na(group)) %>% 
  mutate(label = ifelse(treatment == 'Pre', as.character(patient), '')) %>% 
  pull(label)

p <- total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category)) %>% 
  filter(gene %in% 'IL23A') %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  ggplot(aes(x = sample,
             y = counts,
             fill = group,
             pattern = treatment)) +
  geom_bar_pattern(stat = 'identity', 
                   color = 'black', 
                   alpha = 0.8,
                   pattern_density = 0.1,
                   pattern_angle = 45,
                   pattern_key_scale_factor = 0.6) +
  facet_wrap(~gene, scales = 'free_y', nrow = 4) +
  scale_pattern_manual(values = c(Pre = 'stripe', Mid = 'none')) +
  scale_x_discrete(labels = patient_labels,
                   drop = FALSE) +
  guides(pattern = guide_legend(override.aes = list(fill = 'white')),
         fill = guide_legend(override.aes = list(pattern = 'none'))) +
  scale_fill_manual(values = colors_response2) +
  theme(axis.ticks.x = element_blank(),
        axis.text.x = element_text(hjust = 0),
        legend.position = 'top') +
  labs(y = 'Total UMI / biopsy',
       x = '')

ggsave(plot = p,
	device = cairo_pdf,
	path = output_dir,
        h = 4,
        w = 4,
	filename = 'bar_il23a.pdf')
p
```

### Percentages
```{r}


plot_input <- total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category)) %>%
  filter(gene %in% setdiff(genes_keratinocyte, 'CXCL1')) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  ungroup()

pre_values <- plot_input %>% filter(treatment == 'Pre') %>% select(patient, gene, pre = counts)

p <- plot_input %>% 
  left_join(pre_values) %>% 
  mutate(pct = counts / pre * 100) %>% 
  ggplot(aes(x = sample,
             y = pct,
             fill = category,
             pattern = treatment)) +
  geom_bar_pattern(stat = 'identity', 
                   color = 'black', 
                   alpha = 0.8,
                   pattern_density = 0.1,
                   pattern_angle = 45,
                   pattern_key_scale_factor = 0.6) +
  facet_wrap(~gene, scales = 'free_y', nrow = 4) +
  scale_pattern_manual(values = c(Pre = 'stripe', Mid = 'none')) +
  scale_x_discrete(labels = patient_labels,
                   drop = FALSE) +
  guides(pattern = guide_legend(override.aes = list(fill = 'white')),
         fill = guide_legend(override.aes = list(pattern = 'none'))) +
  scale_fill_manual(values = colors_response3) +
  theme(axis.ticks.x = element_blank(),
        axis.text.x = element_text(hjust = 0),
        legend.position = 'top') +
  labs(y = '% Expression (vs Pre)',
       x = '')

ggsave(plot = p,
       path = output_dir,
       device = cairo_pdf,
       h = 6,
       w = 10,
       filename = paste0('keratinocyte_genes_adjacent_pattern_pct', distance, '.pdf'))
p

```


#### Residual
```{r}
p <-  total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category = group)) %>% 
  filter(gene %in% genes_keratinocyte) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  arrange(gene) %>% 
  group_by(gene, category, treatment) %>% 
  summarize(counts = mean(counts)) %>% 
  pivot_wider(names_from = treatment,
              values_from = counts) %>% 
  mutate(ratio = 100 * Mid/Pre) %>% 
  ggplot(aes(x = category,
             y = ratio,
             fill = category)) +
  geom_boxplot() +
  # geom_col() +
  
  theme_dwu(legend.position = 'top') +
  scale_fill_manual(values = colors_response3) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_x_discrete(drop = FALSE) +
  labs(y = 'Average Residual Expression %',
       x = '')

ggsave(plot = p,
       path = output_dir,
       device = cairo_pdf,
       h = 5,
       w = 2.5,
       filename = paste0('keratinocyte_genes_adjacent', distance, '_mean.pdf'))
p
```
```{r}
p <-  total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category = group)) %>% 
  filter(gene %in% genes_keratinocyte) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  arrange(gene) %>% 
  group_by(gene, category, treatment) %>% 
  summarize(counts = mean(counts)) %>% 
  pivot_wider(names_from = treatment,
              values_from = counts) %>% 
  mutate(ratio = 100 * Mid/Pre) %>% 
  ggplot(aes(x = category,
             y = ratio,
             fill = category)) +
  geom_boxplot() +
  # geom_col() +
  
  theme_dwu(legend.position = 'top') +
  scale_fill_manual(values = colors_response3) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_x_discrete(drop = FALSE) +
  labs(y = 'Average Residual Expression %',
       x = '')

ggsave(plot = p,
       path = output_dir,
       device = cairo_pdf,
       h = 5,
       w = 8,
       filename = paste0('keratinocyte_genes_adjacent', distance, '_mean_by_gene.pdf'))
p
```

```{r}
 total_long_adjacent %>% 
  filter(gene %in% setdiff(genes_keratinocyte, 'CXCL1')) %>% 
  filter(cluster %in% c('Keratinocyte'))
```


```{r}
p <- total_long_adjacent %>% 
  filter(gene %in% genes_keratinocyte) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  ggplot(aes(x = sample,
             y = pct_spots,
             fill = group)) +
  geom_bar(stat = 'summary',
           color = 'black',
           fun = 'sum') +
 # geom_col() +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_x_discrete(drop = FALSE) +
  labs(y = '% Positive Spots',
       x = '')

ggsave(plot = p,
       path = output_dir,
       h = 5,
       w = 18,
       filename = paste0('keratinocyte_genes_pct_adjacent', distance, '.png'))
p
```

```{r}
p <- total_long_adjacent %>% 
  filter(gene %in% genes_il17) %>% 
  filter(cluster == 'Keratinocyte') %>% 
  ggplot(aes(x = sample,
             y = counts,
             fill = group)) +
  geom_bar(stat = 'summary',
           color = 'black',
           fun = 'sum') +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  scale_x_discrete(drop = FALSE) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = 'Total UMI / biopsy',
       x = '')

ggsave(plot = p,
       path = output_dir,
       h = 5,
       w = 14,
       filename = paste0('il17_genes_adjacent', distance, '.png'))
p
```

```{r}
p <- total_long_adjacent %>% 
  filter(gene %in% genes_il17) %>% 
  filter(cluster == 'Keratinocyte') %>% 
  ggplot(aes(x = sample,
             y = pct_spots,
             fill = group)) +
  geom_bar(stat = 'summary',
           color = 'black',
           fun = 'sum') +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = '% Positive Spots',
       x = '')


ggsave(plot = p,
       path = output_dir,
       h = 5,
       w = 14,
       filename = paste0('il17_genes_pct_adjacent', distance, '.png'))
p
```

```{r}
p <- total_long_adjacent %>% 
  filter(gene %in% c('IL17RA', 'IL17RC', 'IL23')) %>% 
  filter(cluster == 'Keratinocyte') %>% 
  ggplot(aes(x = sample,
             y = pct_spots,
             fill = group)) +
  geom_bar(stat = 'summary',
           color = 'black',
           fun = 'sum') +
  facet_wrap(~gene, scales = 'free_y', nrow = 1) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few(drop=FALSE) +
  scale_x_discrete(drop = FALSE) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = '% Positive Spots',
       x = '')


ggsave(plot = p,
       path = output_dir,
       h = 4,
       w = 5,
       filename = paste0('keratinocytes_il17_genes_pct_adjacent', distance, '.png'))
p
```
### Stromal 
```{r}
p <- total_long_adjacent %>% 
  filter(gene %in% genes_stromal) %>% 
  filter(cluster %in% c('Stromal', 'Fibroblast')) %>% 
  ggplot(aes(x = sample,
             y = pct_spots,
             fill = group)) +
  geom_bar(stat = 'summary',
           color = 'black',
           fun = 'sum') +
 # geom_col() +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme_dwu(legend.position = 'top') +
  ggthemes::scale_fill_few() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_x_discrete(drop = FALSE) +
  labs(y = '% Positive Spots',
       x = '')

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 5,
       w = 18,
       filename = paste0('stromal_genes_pct_adjacent_stromal', distance, '.pdf'))
p
```

### Stats
```{r}
stat_input <- total_long_adjacent %>% 
  left_join(patient_metadata %>% select(id, category)) %>% 
  filter(gene %in% genes_keratinocyte) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  filter(treatment == 'Mid') %>% 
  arrange(gene) %>% 
  group_by(gene, category) %>% 
  summarize(counts = mean(counts))

stat_input


```
```{r}

```

```{r}
group1 <- stat_input %>% filter(category == 'Responder') %>% pull(counts)
group2 <- stat_input %>% filter(category == 'Non-responder (Trm-insensitive)') %>% pull(counts)
wilcox.test(group1, group2, paired = TRUE)
```
```{r}
stat_input <- total_long_adjacent %>% 
  filter(gene %in% genes_keratinocyte) %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  arrange(gene) %>% 
  group_by(gene, group, treatment) %>% 
  summarize(counts = mean(counts)) %>% 
  pivot_wider(names_from = treatment,
              values_from = counts) %>% 
  mutate(ratio = Mid/Pre)

stat_input
```


```{r}
stat_input %>% drop_na() %>% 
  group_by(group) %>% 
  summarize(median = median(ratio))
```

### Radial analysis

```{r}
th17_spots <- counts_table %>% filter(spatial_CD3D + spatial_CD4 + spatial_CD8A + spatial_CD8B > 0, 
                                      spatial_IL17A + spatial_IL17F > 0) %>% 
  pull(rowname)
th17_spots
```


```{r}
distance_to_th17 <- th17_spots %>% 
  map(function(i) {
    spatial_neighbors(seuratobj = seuratobj,
                      distance = 10,
                      barcode = i) 
  }) %>% bind_rows() %>% 
  group_by(rowname, id, sample, patient, treatment, group) %>% 
  slice_min(distance, n = 1) %>% 
  ungroup() 

distance_to_th17
```

```{r}
counts_distance <- counts_table %>% 
  filter(cluster %in% c('Keratinocyte')) %>% 
  #filter(cluster %in% c('Stromal', 'Fibroblast')) %>% 
  left_join(distance_to_th17 %>% 
              select(rowname, distance, treatment, group) %>% unique()) %>% 
  pivot_longer(cols = starts_with('spatial'),
               names_to = 'gene',
               values_to = 'count') %>% 
  mutate(gene = str_remove(gene, 'spatial_'))

counts_distance
```
```{r}
counts_distance_summarized <-  counts_distance %>% 
    filter(distance != 0) %>% 
    group_by(gene, group, treatment, distance) %>% 
    summarize(count = sum(count),
              nCount_Spatial = sum(nCount_Spatial)) %>% 
  mutate(cpm = count / nCount_Spatial * 1e6) %>% 
  drop_na() 

p <- map(counts_distance_summarized$gene %>% unique() %>% intersect(c(genes_keratinocyte) %>% setdiff('CXCL1')), function(i) {
#p <- map(counts_distance_summarized$gene %>% unique() %>% intersect(genes_control), function(i) {
  
  print2(i)
  counts_distance_summarized %>% 
    filter(gene == i) %>% 
    ggplot(aes(x = distance,
               y = cpm,
               color = factor(treatment, levels = c('Pre', 'Mid')),
               group = treatment)) +
    geom_point(size = 2, 
               alpha = 0.8) +
    geom_path() +
    facet_wrap(~factor(group, levels = c('Responder', 'Non-responder')), ncol = 2) +
    labs(title = i,
         x = 'Distance to T17 infiltrate (# spots)',
         y = 'Counts per million') +
    scale_color_manual(values = c('orangered', 'dodgerblue4')) +
    theme(legend.position = 'top',
          plot.background = element_rect(color = 'black')) +
    scale_x_continuous(breaks = 1:10)
}) %>% wrap_plots()

p

```
```{r}
p <- map(counts_distance_summarized$gene %>% unique() %>% intersect(c(genes_keratinocyte) %>% setdiff('CXCL1')), function(i) {
#p <- map(counts_distance_summarized$gene %>% unique() %>% intersect(genes_control), function(i) {
  
  print2(i)
  counts_distance_summarized %>% 
    filter(gene == i) %>% 
    ggplot(aes(x = distance,
               y = count,
               color = factor(treatment, levels = c('Pre', 'Mid')),
               group = treatment)) +
    geom_point(size = 2, 
               alpha = 0.8) +
    geom_path() +
    facet_wrap(~factor(group, levels = c('Responder', 'Non-responder')), ncol = 2) +
    labs(title = i,
         x = 'Distance to T17 infiltrate (# spots)',
         y = 'Total UMI') +
    scale_color_manual(values = c('orangered', 'dodgerblue4')) +
    theme(legend.position = 'top',
          plot.background = element_rect(color = 'black')) +
    scale_x_continuous(breaks = 1:10)
}) %>% wrap_plots()

p
```
#### Line plots
```{r}
counts_distance_summarized <-  counts_distance %>% 
    filter(distance != 0) %>% 
  
    group_by(gene, group, treatment, distance) %>% 
    summarize(count = sum(count)) %>% 
  drop_na() 

p <- map(counts_distance_summarized$gene %>% unique() %>% intersect(genes_keratinocyte), function(i) {
  
  print2(i)
  counts_distance_summarized %>% 
    filter(gene == i) %>% 
    ggplot(aes(x = distance,
               y = count,
               color = factor(treatment, levels = c('Pre', 'Mid')),
               group = treatment)) +
    geom_point(size = 2, 
               alpha = 0.8) +
    geom_path() +
    facet_wrap(~factor(group, levels = c('Responder', 'Non-responder')), ncol = 2) +
    labs(title = i,
         x = 'Distance to immune infiltrate (# spots)',
         y = 'Total UMI') +
    scale_color_manual(values = c('orangered', 'dodgerblue4')) +
    theme(legend.position = 'top') +
    scale_x_continuous(breaks = 1:10)
}) %>% wrap_plots()

p
```
#### Line plots subset
```{r}
gene_subset <- c('DEFB4A', 
                 'IL36G',
                 'LCN2',
                 'PI3',
                 'RHCG',
                 'S100A7A',
                 'SERPINB3',
                 'SLC6A14',
                 'SPRR2A')

counts_distance_summarized <-  counts_distance %>% 
    filter(distance != 0) %>% 
    group_by(gene, distance) %>% 
    summarize(count = sum(count),
              nCount_Spatial = sum(nCount_Spatial)) %>% 
  mutate(cpm = count / nCount_Spatial * 1e6) %>% 
  drop_na() 

p <- map(counts_distance_summarized$gene %>% intersect(gene_subset), function(i) {
  
  print2(i)
  counts_distance_summarized %>% 
    filter(gene == i) %>% 
    ggplot(aes(x = distance,
               y = cpm)) +
    geom_point(size = 2, 
               alpha = 0.8) +
    geom_path() +
    labs(title = i,
         x = 'Distance to T17 infiltrate (# spots)',
         y = 'Counts per million') +
    scale_color_manual(values = c('orangered', 'dodgerblue4')) +
    theme(legend.position = 'top',
          plot.background = element_rect(color = 'black')) +
    scale_x_continuous(breaks = 1:10)
}) %>% wrap_plots()

p
```
### Spatial diagram
#### Combined
```{r}
p <- counts_distance_summarized %>% 
  group_by(gene) %>% 
  mutate(max = max(cpm),
         pct = cpm / max * 100) %>% 
  filter(gene %in% gene_subset) %>% 
  ggplot(aes(x = gene,
             y = distance,
             fill = pct)) + 
  geom_point(shape = 21, aes(size = distance)) +
  ylim(c(0, 5)) +
  theme_void() +
  theme(legend.position = 'right',
        axis.text.x = element_text()) + 
  scale_size_continuous(range = c(10, 10), guide = NULL) +
  scale_fill_gradientn(colors = colorRampPalette(RColorBrewer::brewer.pal(n = 7, name = "OrRd"))(100),
                       limits = c(0, 100),
                       breaks = seq(0, 100, length.out = 5),
                       guide = guide_colorbar(title = '% of max', ticks = FALSE, frame.colour = 'black')) +
  coord_polar(theta = 'x')

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 5,
       w = 5,
       filename = 'polar_expression.pdf')

p
```
#### Th1 and Th2 genes
```{r}
gene_subset <- genes_th1

p <- map(counts_distance_summarized$gene %>% intersect(gene_subset), function(i) {
  
  print2(i)
  counts_distance_summarized %>% 
    filter(gene == i) %>% 
    ggplot(aes(x = distance,
               y = cpm)) +
    geom_point(size = 2, 
               alpha = 0.8) +
    geom_path() +
    labs(title = i,
         x = 'Distance to T17 infiltrate (# spots)',
         y = 'Counts per million') +
    scale_color_manual(values = c('orangered', 'dodgerblue4')) +
    theme(legend.position = 'top',
          plot.background = element_rect(color = 'black'),
          plot.title = element_text(face = 'italic')) +
    scale_x_continuous(breaks = 1:10)
}) %>% wrap_plots()

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 8,
       w = 12,
       filename = 'distance_th1.pdf')

p
```
#### Th1 and Th2 genes
```{r}
gene_subset <- genes_th2

p <- map(counts_distance_summarized$gene %>% intersect(gene_subset), function(i) {
  
  print2(i)
  counts_distance_summarized %>% 
    filter(gene == i) %>% 
    ggplot(aes(x = distance,
               y = cpm)) +
    geom_point(size = 2, 
               alpha = 0.8) +
    geom_path() +
    labs(title = i,
         x = 'Distance to T17 infiltrate (# spots)',
         y = 'Counts per million') +
    scale_color_manual(values = c('orangered', 'dodgerblue4')) +
    theme(legend.position = 'top',
          plot.background = element_rect(color = 'black'),
          plot.title = element_text(face = 'italic')) +
    scale_x_continuous(breaks = 1:10)
}) %>% wrap_plots()

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 8,
       w = 12,
       filename = 'distance_th2.pdf')

p
```
```{r}
gene_subset <- genes_keratinocyte

p <- map(counts_distance_summarized$gene %>% intersect(gene_subset), function(i) {
  
  print2(i)
  counts_distance_summarized %>% 
    filter(gene == i) %>% 
    ggplot(aes(x = distance,
               y = cpm)) +
    geom_point(size = 2, 
               alpha = 0.8) +
    geom_path() +
    labs(title = i,
         x = 'Distance to T17 infiltrate (# spots)',
         y = 'Counts per million') +
    scale_color_manual(values = c('orangered', 'dodgerblue4')) +
    theme(legend.position = 'top',
          plot.background = element_rect(color = 'black'),
          plot.title = element_text(face = 'italic')) +
    scale_x_continuous(breaks = 1:10)
}) %>% wrap_plots()

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 8,
       w = 12,
       filename = 'distance_keratinocyte_il17.pdf')
p
```


### Sector plot
```{r}
p <- counts_distance_summarized %>% 
  filter(distance %in% 1:5) %>% 
  group_by(gene) %>% 
  mutate(max = max(cpm),
         pct = cpm / max * 100) %>% 
  filter(gene %in% gene_subset) %>% 
  ggplot(aes(x = gene,
             y = distance/distance,
             fill = pct)) + 
  geom_bar(stat = 'identity',
           color = 'black',
           position = 'stack') +
  theme_void() +
  theme(legend.position = 'right',
        axis.text.x = element_text()) + 
  scale_fill_gradientn(colors = colorRampPalette(RColorBrewer::brewer.pal(n = 7, name = "OrRd"))(100),
                       limits = c(0, 100),
                       breaks = seq(0, 100, length.out = 5),
                       guide = guide_colorbar(title = '% of max', ticks = FALSE, frame.colour = 'black')) +
  coord_polar(theta = 'x')

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 5,
       w = 6,
       filename = 'polar_sectors.pdf')

p

```


### Old
```{r}
test_spatial %>% group_by(id) %>% tally()
```

```{r}
test_spatial %>% 
  filter(id == 'S296') %>% 
  select(imagerow, imagecol) %>% 
  summarize_all(median)
```
```{r}
test_spatial %>% 
  filter(id == 'S296',
         abs(imagerow - 12700) < 100,
         abs(imagecol - 13135) < 100)
```


```{r}
barcode <- 'GTGTTACTATGCGTCC-1_8'
neighbors <- spatial_neighbors(seuratobj,
                               barcode = barcode,
                               distance = 5)

neighbors
```

```{r}
neighbors %>% 
  ggplot(aes(x = imagerow,
             y = imagecol,
             fill = factor(distance))) +
  geom_point(shape = 21,
             size = 10) + 
  ggthemes::scale_fill_tableau(palette = 'Tableau 10') +
  coord_equal() +
  theme(legend.position = 'right')
```
```{r}
distance_map <- neighbors %>% select(imagerow, imagecol, distance)
distance_map
```
```{r}
p <- map(c('IL17A', 'IL17F', 'IFNG', 'IL26'), function(i) {
  
counts_distance_summarized %>% 
  filter(gene == i) %>% 
  left_join(distance_map) %>% 
    mutate(group = factor(group, levels = c('Responder', 'Non-responder'))) %>% 
    mutate(treatment = factor(treatment, levels = c('Pre', 'Mid'))) %>% 
  ggplot(aes(x = imagerow,
             y = imagecol,
             fill = count)) + 
   geom_point(shape = 21,
              size = 4) + 
    coord_equal() +
    #facet_wrap(~factor(treatment, levels = c('Pre', 'Mid'))) +
    facet_wrap(~group + treatment) +
    labs(title = i,
         x = NULL,
         y = NULL) +
    scale_fill_gradientn(colors = colorRampPalette(RColorBrewer::brewer.pal(n = 7, name = "OrRd"))(100),
                        guide = guide_colorbar(title = 'Total UMI', ticks = FALSE, frame.colour = 'black')) +
    theme(legend.position = 'right',
          legend.title = element_text(hjust = 0.5),
          axis.text = element_blank(),
          axis.ticks = element_blank()) 
  
}) %>% wrap_plots()

p

```

### loop for all genes
```{r}
spots_per_gene <- rowSums(seuratobj@assays$Spatial@counts > 0) %>% 
  sort(decreasing = TRUE) %>% 
  as.data.frame() %>% 
  set_names('spots') %>% 
  rownames_to_column('gene') %>% 
  as_tibble()

spots_per_gene %>% 
  ggplot(aes(x = spots)) +
  geom_histogram(bins = 100) +
  scale_x_log10() +
  geom_vline(xintercept = 500)
```
```{r}
spots_per_gene %>% filter(spots > 0) %>% summarize(mean = mean(spots),
                                                   sd = sd(spots),
                                                   median = median(spots))
```
```{r}
genes_to_check <- spots_per_gene %>% filter(spots > median(spots)) %>% pull(gene)
```


```{r}
distance_cor <- map(genes_to_check, function(i) {
  
  print2(i)
  
  counts_table <- FetchData(seuratobj,
                          vars = c('id', 'sample', 'ident', 'nCount_Spatial', paste0('spatial_', i) %>% unique), 
                          slot = 'counts') %>% 
  rownames_to_column() %>% 
  as_tibble()

  counts_distance <- counts_table %>% 
    left_join(distance_to_th17 %>% 
                select(rowname, distance, treatment, group) %>% unique()) %>% 
    pivot_longer(cols = starts_with('spatial'),
                 names_to = 'gene',
                 values_to = 'count') %>% 
    mutate(gene = str_remove(gene, 'spatial_'))
  
  counts_distance_summarized <- counts_distance %>% 
    filter(distance != 0) %>% 
    group_by(group, treatment, distance) %>% 
    summarize(count = sum(count)) %>% 
    drop_na() 
  
  rho <- cor(counts_distance_summarized$distance, counts_distance_summarized$count, method = 'spearman')
  
  # p <- counts_distance_summarized %>% 
  #   ggplot(aes(x = distance,
  #              y = count,
  #              color = factor(treatment, levels = c('Pre', 'Mid')),
  #              group = treatment)) +
  #   geom_point(size = 2, 
  #              alpha = 0.8) +
  #   geom_path() +
  #   facet_wrap(~factor(group, levels = c('Responder', 'Non-responder')), ncol = 2) +
  #   labs(title = i,
  #        x = 'Distance to immune infiltrate (# spots)',
  #        y = 'Total UMI') +
  #   scale_color_manual(values = c('orangered', 'dodgerblue4')) +
  #   theme(legend.position = 'top') +
  #   scale_x_continuous(breaks = 1:5)
  
  
  
  tibble('gene' = i,
         'rho' = round(rho, 3))
  
}) %>% bind_rows()

```

```{r}

```
```{r}



p <- map(counts_distance_summarized$gene %>% unique() %>% intersect(genes_control), function(i) {
  
  
}) %>% wrap_plots()

p

```

```{r}
counts_distance_summarized %>% 
  write_tsv(here(output_dir, 'counts_by_distance.tsv'))
  
```
