==> analysis/scripts/12_cytokines.Rmd <==
---
title: "Origin of Cytokines"
author: "David Wu"
---

## Setup
### Use the here package for robust relative paths
```{r}
library(here)
```

### Optionally convert this Rmd to R script 
```{r eval=FALSE, message=FALSE}
current_file <- rstudioapi::getActiveDocumentContext()$path
output_file <- stringr::str_replace(current_file, '.Rmd', '.R')
knitr::purl(current_file, output = output_file)
file.edit(output_file)
```

### Directories
```{r, message=FALSE}
output_dir <- here('analysis/output/11_cytokines') # analysis file output directory
data_dir <- here('data/derived/tildra') # data file output directory

dir.create(output_dir, showWarnings = FALSE)
dir.create(data_dir, showWarnings = FALSE)
```

### Loading the libraries
```{r, message=FALSE}
library(Seurat) 
library(tidyverse)
library(patchwork)
library(SeuratWrappers) # devtools::install_github('satijalab/seurat-wrappers')
library(wutilities) # devtools::install_github('symbiologist/wutilities')
library(tictoc)
library(extrafont)
loadfonts()
theme_set(theme_dwu()) # set default theme
```



### Import
```{r}
seuratobj <- read_rds(here(data_dir, 'seuratobj_annotated.rds'))
seuratobj
```
### Use global or local identities and set as cluster/supercluster in the object
```{r}
which_identity <- 'local'
which_cluster <- paste0(which_identity, '_cluster')
which_supercluster <- paste0(which_identity, '_supercluster')
which_supercluster <- paste0(which_identity, '_supercluster')

seuratobj$cluster <- seuratobj[[which_cluster]] 
seuratobj$supercluster <- seuratobj[[which_supercluster]] 
seuratobj$supercluster <- seuratobj[[which_supercluster]] 
Idents(seuratobj) <- 'supercluster' # set cluster as main annotation

output_dir <- here(output_dir, which_identity)
dir.create(output_dir)
```

### Prepare metadata subset
```{r}
cytokines <- c("IFNG",
               "TNF",
               "IL17A",
               "IL17F",
               "IL21",
               "IL22",
               "IL26",
               #"IL36A",
               #'IL36B',
               'IL36G',
               "CXCL13")

cells <- seuratobj@meta.data %>% filter(condition != 'AD') %>% rownames()

pseudo <- pseudobulk(seuratobj,
                     genes = cytokines,
                     cells = cells,
                     group_by = c('patient', 'supercluster', 'condition', 'treatment', 'group'))

pseudo %>% head()
```

```{r}
sample_total <- pseudo %>% 
  group_by(patient, condition, group, treatment) %>% 
  summarize(sample_total = sum(total))
sample_total
```
```{r}
cytokine_total <- pseudo %>% 
  group_by(patient, treatment, gene) %>% 
  summarize(cytokine_total = sum(counts))

cytokine_total
```


```{r}
cytokine_source <- pseudo %>% 
  left_join(cytokine_total) %>% 
  left_join(sample_total) %>% 
  mutate(cytokine_pct = round(100 * counts / cytokine_total, 2),
         sample_cpm = round(counts / sample_total * 1e6, 3))

gene_counts <- cytokine_source %>% 
  group_by(gene, patient, treatment) %>% 
  summarize(gene_total = sum(counts))

cytokine_source <- cytokine_source %>% left_join(gene_counts)
cytokine_source
```
```{r}
which_patient <- 'P2'
which_treatment <- 'Pre'

annotation <- levels(cytokine_source$supercluster)
color_scale <- c(RColorBrewer::brewer.pal(n = sum(str_detect(annotation, 'T|B|ILC')),
                                          name = 'Blues'), 'black',
                 RColorBrewer::brewer.pal(n = sum(!str_detect(annotation, 'T|B|ILC')),
                                          name = 'Reds'))
                 

cytokine_source %>% 
  filter(patient == which_patient,
         treatment == which_treatment) %>% 
  ggplot(aes(x = gene,
             y = cytokine_pct,
             fill = supercluster)) +
  geom_col(color = 'black') +
  labs(x = 'Cytokine',
       y = '% Produced by Cell Type',
       title = paste('Patient', which_patient)) +
  theme(legend.position = 'right') +
  #ggthemes::scale_fill_tableau(palette = 'Tableau 20') +
  scale_fill_manual(values = color_scale)


```

### Loop
```{r}
tildra_patients <- cytokine_source %>% filter(treatment != 'None') %>% pull(patient) %>% unique() %>% sort()

p <- map(tildra_patients %>% unique(), function(which_patient) {

  p_pre <- cytokine_source %>% 
    filter(patient == which_patient,
           treatment == 'Pre') %>% 
    ggplot(aes(x = gene,
               y = cytokine_pct,
               fill = supercluster)) +
    geom_col(color = 'black') +
    labs(x = 'Cytokine',
         y = '% Produced by Cell Type',
         title = paste(which_patient, 'Pre')) +
    theme(legend.position = 'none') +
    scale_fill_manual(values = color_scale)
  
  p_mid <- cytokine_source %>% 
    filter(patient == which_patient,
           treatment == 'Mid') %>% 
    ggplot(aes(x = gene,
               y = cytokine_pct,
               fill = supercluster)) +
    geom_col(color = 'black') +
    labs(x = 'Cytokine',
         y = '% Produced by Cell Type',
         title = paste(which_patient, 'Mid')) +
    theme(legend.position = 'right') +
    scale_fill_manual(values = color_scale)
  
  p <- p_pre + p_mid
  
}) %>% wrap_plots()


```
#### Facets
```{r}
tildra_patients <- cytokine_source %>% filter(treatment != 'None') %>% pull(patient) %>% unique() %>% sort()

p <- map(tildra_patients %>% unique(), function(which_patient) {

  cytokine_source %>% 
    filter(patient == which_patient) %>% 
    ggplot(aes(x = gene,
               y = cytokine_pct)) +
    geom_col(color = 'black', aes(fill = supercluster)) +
    geom_text(data = cytokine_source %>% filter(patient == which_patient) %>% 
                select(gene, treatment, gene_total) %>% unique(),
              aes(x = gene,
                  y = 106,
                  label = gene_total),
              size = 3) +
    labs(x = 'Cytokine',
         y = '% Produced by Cell Type',
         title = paste(which_patient)) +
    facet_wrap(~treatment, nrow = 2) +
    theme(legend.position = 'right',
          axis.text.x = element_text(angle = 45, hjust = 1),
          plot.background = element_rect(color = 'black')) +
    scale_fill_manual(values = color_scale) +
    scale_y_continuous(breaks = c(0, 25, 50, 75, 100))
  
}) %>% wrap_plots()
```


### Absolute scale
```{r}
cytokine_source %>% 
  filter(patient == which_patient, treatment == 'Pre') %>% 
  ggplot(aes(x = gene,
             y = cpm,
             fill = supercluster)) +
  geom_bar(color = 'black', stat = 'identity') +
  # geom_text(data = cytokine_source %>% filter(patient == which_patient) %>% select(treatment, gene, cpm, gene_total) %>% unique(),
  #           aes(x = gene, y = cpm + 1, label = gene_total),
  #           inherit.aes = FALSE) +
  labs(x = 'Cytokine',
       y = 'Counts per Million by Cell Type',
       title = paste('Patient', which_patient)) +
  theme(legend.position = 'right') +
  scale_fill_manual(values = color_scale)
```
```{r}
cytokine_source %>% 
  filter(patient == which_patient, treatment == 'Pre') %>% 
  ggplot(aes(x = gene,
             y = cpm + 1,
             fill = supercluster)) +
  geom_bar(color = 'black', stat = 'identity') +
  # geom_text(data = cytokine_source %>% filter(patient == which_patient) %>% select(treatment, gene, cpm, gene_total) %>% unique(),
  #           aes(x = gene, y = cpm + 1, label = gene_total),
  #           inherit.aes = FALSE) +
  labs(x = 'Cytokine',
       y = 'Counts per Million by Cell Type',
       title = paste('Patient', which_patient)) +
  theme(legend.position = 'right') +
  scale_fill_manual(values = color_scale) 
```

```{r}
cytokine_source %>% filter(patient == which_patient, gene == 'CXCL13', treatment == 'Pre') %>% pull(cpm)
```

### Group
```{r}
cytokine_source %>% 
  ggplot(aes(x = supercluster,
             y = cytokine_pct)) +
  geom_boxplot() +
  facet_wrap(~gene)
```
```{r}
cytokine_source %>% filter(patient == which_patient) %>% select(treatment, gene, cpm, gene_total)

```

### As heatmap
```{r}
cytokine_source_combined <- pseudo %>% 
  left_join(cytokine_total) %>% 
  left_join(sample_total) %>% 
  group_by(supercluster, gene, condition, treatment, group) %>% 
  summarize(counts = sum(counts),
            total = sum(total),
            cytokine_total = sum(cytokine_total)) %>% 
  mutate(cytokine_pct = round(100 * counts / cytokine_total, 2))

gene_counts <- cytokine_source %>% 
  group_by(gene, condition, group, treatment) %>% 
  summarize(gene_total = sum(counts))

cytokine_source_combined <- cytokine_source_combined %>% left_join(gene_counts)
cytokine_source_combined
```

```{r}
heatmap_input <- cytokine_source_combined %>% 
  select(supercluster, gene, condition, treatment, group, counts, total, cytokine_pct) %>% 
  unique() %>% 
  mutate(description = paste(condition, group, treatment)) %>% 
  mutate(description = description %>% str_remove(' NA None') %>% str_remove('PV ')) %>% 
  mutate(description = factor(description, levels =
                                c('PV', 
                                'Non-responder Pre',
                                'Non-responder Mid',
                                'Responder Pre',
                                'Responder Mid',
                                'HC'))) %>% 
  arrange(condition, treatment, group, supercluster, cytokine_pct)

heatmap_input
```
```{r}
color_scale <- c(0, 50, 100)
cluster_order <- heatmap_input %>% filter(description == 'PV', gene == 'IL17A') %>% arrange(cytokine_pct) %>% pull(supercluster) %>% unique()
  
p <- heatmap_input %>% 
  filter(gene %in% c('IL17A', 'IL17F', 'IL26', 'IFNG')) %>% 
  mutate(gene = factor(gene, levels = c('IL17A', 'IL17F', 'IL26', 'IFNG'))) %>% 
  mutate(cytokine_pct = round(cytokine_pct, 1)) %>% 
  ggplot(aes(x = description,
             y = factor(supercluster, levels = cluster_order),
             fill = cytokine_pct,
             label = cytokine_pct)) +
  geom_tile(color = 'black') +
  geom_text(aes(color = cytokine_pct), size = 3) +
  scale_color_gradient(low = 'dodgerblue4',
                       high = 'white',
                       guide = NULL) +
  scale_fill_gradientn(colors = colorRampPalette(RColorBrewer::brewer.pal(n = 7, name = "Blues"))(100),
                         limits = range(color_scale),
                         breaks = color_scale,
                         oob = scales::squish,
                         na.value = 'grey90',
                         guide = guide_colorbar(frame.colour = 'black', 
                                                ticks = FALSE, 
                                                nbin = 100,
                                                barwidth = 0.5,
                                                title = '%',
                                                barheight = 5)) +
  theme(legend.position = 'right',
        axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1),
        axis.ticks = element_blank(),
        axis.line = element_blank(),
        panel.border = element_blank(),
        legend.title = element_text(hjust = 0.5)) +
  facet_wrap(~gene, nrow = 1) +
  labs(x = NULL,
       y = NULL,
       title = 'Cytokine Source')

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 6,
        w = 9,
	filename = 'F4B_cytokine_source.pdf')

p
```
```{r}
color_scale <- c(0, 2500, 5000)
cluster_order <- heatmap_input %>% filter(description == 'PV', gene == 'IL17A') %>% arrange(cytokine_pct) %>% pull(supercluster) %>% unique()
  
p <- heatmap_input %>% 
  filter(gene %in% c('IL17A', 'IL17F', 'IL26', 'IFNG')) %>% 
  mutate(gene = factor(gene, levels = c('IL17A', 'IL17F', 'IL26', 'IFNG'))) %>% 
  ggplot(aes(x = description,
             y = factor(supercluster, levels = cluster_order),
             fill = counts,
             label = counts)) +
  geom_tile(color = 'black') +
  geom_text(aes(color = counts), size = 3) +
  scale_color_gradient(low = 'dodgerblue4',
                       high = 'white',
                       guide = NULL) +
  scale_fill_gradientn(colors = colorRampPalette(RColorBrewer::brewer.pal(n = 7, name = "Blues"))(100),
                         limits = range(color_scale),
                         breaks = color_scale,
                         oob = scales::squish,
                         na.value = 'grey90',
                         guide = guide_colorbar(frame.colour = 'black', 
                                                ticks = FALSE, 
                                                nbin = 100,
                                                barwidth = 0.5,
                                                title = '%',
                                                barheight = 5)) +
  theme(legend.position = 'right',
        axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1),
        axis.ticks = element_blank(),
        axis.line = element_blank(),
        panel.border = element_blank(),
        legend.title = element_text(hjust = 0.5)) +
  facet_wrap(~gene, nrow = 1) +
  labs(x = NULL,
       y = NULL,
       title = 'Cytokine Source')

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 6,
        w = 9,
	filename = 'F4B_cytokine_source_abs.pdf')

p
```

### Per cell normalized
```{r}
cells_per_sample <- seuratobj@meta.data %>% 
  select(patient, treatment, group, supercluster) %>% 
  group_by_all() %>% 
  tally()

cells_per_sample
```
```{r}
per_cell_input <- pseudo %>% 
  filter(gene %in% c('IL17A', 'IL17F', 'IL26', 'IFNG')) %>% 
  filter(supercluster == 'Trm') %>% 
  left_join(cells_per_sample) %>% 
    mutate(description = paste(condition, group, treatment)) %>% 
  mutate(description = description %>% str_remove(' NA None') %>% str_remove('PV ')) %>% 
  mutate(description = factor(description, levels =
                                c('PV', 
                                'Non-responder Pre',
                                'Non-responder Mid',
                                'Responder Pre',
                                'Responder Mid',
                                'HC'))) %>% 
  mutate(counts_per_cell = counts/n,
         cpm_per_cell = cpm/n) %>% 
  unique()

per_cell_input
  
```
```{r}
p <- per_cell_input %>% 
  ggplot(aes(x = n,
             y = counts,
             color = description)) +
  geom_point() +
  facet_wrap(~gene,  scales = 'free', nrow = 1) +
  scale_x_log10() + 
  scale_y_log10() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = 'right') +
  labs(x = '# Trm Cells',
       y = 'Cytokine UMIs')
p
```
```{r}
p <- per_cell_input %>% 
  ggplot(aes(x = description,
             y = counts_per_cell,
             fill = description)) +
  geom_boxplot() +
  facet_wrap(~gene,  scales = 'free_y', nrow = 1) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(x = NULL,
       y = 'Cytokine UMIs per Trm cell')

ggsave(plot = p,
	device = cairo_pdf,
	path = output_dir,
        h = 3.5,
        w = 7,
	filename = 'F4B_cytokines_per_cell.pdf')
p
```

### stats

```{r}
per_cell_input
```

```{r}
p <- per_cell_input %>% 
  ggplot(aes(x = n,
             y = counts,
             color = description)) +
  geom_point() +
  facet_wrap(~gene,  scales = 'free', nrow = 1) +
  scale_x_log10() + 
  scale_y_log10() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = 'right') +
  labs(x = '# Trm Cells',
       y = 'Cytokine UMIs')
p
```

## Session Info
```{r}
sessionInfo()
```
