==> analysis/scripts/12_il23_vs_il17.Rmd <==
---
title: "Tildrakizumab vs Secukinumab analysis"
author: "David Wu"
---

Set working directory to project directory
```{r setup}
library(here)
```

### Optionally convert this Rmd to R script 
```{r eval=FALSE, message=FALSE}
current_file <- rstudioapi::getActiveDocumentContext()$path
output_file <- stringr::str_replace(current_file, '.Rmd', '.R')
knitr::purl(current_file, output = output_file)
file.edit(output_file)
```

### Directories
```{r, message=FALSE}
output_dir <- here('analysis/output/10_il23_vs_il17') # analysis file output directory
data_dir <- here('data/derived/tildra') # data file output directory

dir.create(output_dir, showWarnings = FALSE)
dir.create(data_dir, showWarnings = FALSE)
```

### Loading the libraries
```{r, message=FALSE}
library(Seurat) 
library(tidyverse)
library(ggExtra)
library(ggthemes)
library(ggforce)
library(wutilities) # devtools::install_github('symbiologist/wutilities')
library(tictoc)
library(rcartocolor)
library(ggrepel)
library(DESeq2)
library(edgeR)
theme_set(theme_dwu()) # set default theme
```


---
title: "DE analysis"
author: "David Wu"
---

## Setup
### Use the here package for robust relative paths
```{r}
library(here)
```

### Optionally convert this Rmd to R script 
```{r eval=FALSE, message=FALSE}
current_file <- rstudioapi::getActiveDocumentContext()$path
output_file <- stringr::str_replace(current_file, '.Rmd', '.R')
knitr::purl(current_file, output = output_file)
file.edit(output_file)
```

### Directories
```{r, message=FALSE}
output_dir <- here('analysis/output/08_de_pseudobulk') # analysis file output directory
data_dir <- here('data/derived/tildra') # data file output directory

dir.create(output_dir, showWarnings = FALSE)
dir.create(data_dir, showWarnings = FALSE)
```

### Loading the libraries
```{r, message=FALSE}
library(Seurat) 
library(tidyverse)
library(ggExtra)
library(SeuratWrappers) # devtools::install_github('satijalab/seurat-wrappers')
library(wutilities) # devtools::install_github('symbiologist/wutilities')
library(tictoc)
library(edgeR)
library(openxlsx)
```

### Custom functions
```{r}
aggregate_matrix <- function(input_matrix,
                             rows = NULL) {
  input_matrix <- as.matrix(input_matrix)
  aggregated_matrix <- t(sapply(by(input_matrix, rows, colSums), identity))
}



de_pseudobulk_pipeline <- function(pseudobulk_subset,
                                   metadata_subset,
                                   cluster = 'All',
                                   comparison = NULL,
                                   paired = FALSE,
                                   group1_id,
                                   group2_id) {
  
  comparison <- paste(group1_id, 'vs', group2_id) 

  metadata_table <- metadata_subset %>% 
    filter(id %in% colnames(pseudobulk_subset)) %>% 
    droplevels() %>% 
    arrange(id, colnames(pseudobulk_subset))
  
  if(paired) {
    design <- model.matrix(data = metadata_table, ~patient + group)  
  } else {
    design <- model.matrix(data = metadata_table, ~group)  
  }
  
  
  dge <- DGEList(counts = pseudobulk_subset,
                 group = metadata_table$group)
  
  #keep <- filterByExpr(dge, min.count = 0, min.prop = 0)
  
  dge <- dge %>% calcNormFactors(method = 'TMM')
  dge <- dge %>% estimateDisp(design = design)
  fit <- dge %>% glmFit(design = design)
  res <- fit %>% glmLRT()
  
  de_table <- topTags(res, n = Inf) %>% 
    as.data.frame() %>% 
    rownames_to_column('gene') %>% 
    as_tibble() %>% 
    mutate(logFC = -1 * logFC, # reverse sign such that group 1 is higher
           cluster = cluster,
           comparison = comparison,
           group1 = group1_id,
           group2 = group2_id,
           n_group1 = metadata_table %>% filter(group == 'group1') %>% nrow(),
           n_group2 = metadata_table %>% filter(group == 'group2') %>% nrow()) %>% 
    select(comparison, cluster, gene, group1, group2, n_group1, n_group2,  avg_log2FC = logFC, p_val = PValue, p_val_adj = FDR, everything())
}

write_excel_notebook <- function(input_table, 
                                 output_file, 
                                 tabs = 'cluster', 
                                 prefix = 'cluster_') {
  OUT <- createWorkbook()
  
  input_table$tabs <- input_table[[tabs]]
  tabs <- input_table[[tabs]] %>% unique()
  
  
  for(i in tabs){
    tryCatch({
      table_subset <- input_table %>% filter(tabs == i) %>% select(-tabs)
      
      sheet_name <- paste0(prefix, i) %>% str_remove('[\\?/]') %>% str_remove('[\\/]') %>% str_remove('[\\:/]') 
      addWorksheet(OUT, sheet_name)
      writeData(OUT, sheet = sheet_name, x = table_subset)
      
    }, error=function(e){message(paste0("We have an error here in ", i))})
  }
  
  saveWorkbook(OUT, output_file, overwrite = TRUE)
}
```

### Import
```{r}
seuratobj <- read_rds(here(data_dir, 'seuratobj_annotated.rds'))
seuratobj
```
### Set identity
```{r}
which_annotation <- 'local_supercluster'
seuratobj$annotation <- seuratobj[[which_annotation]]
Idents(seuratobj) <- 'annotation'

metadata <- seuratobj@meta.data %>% rownames_to_column()

output_dir <- here(output_dir, which_annotation)
dir.create(output_dir)
```

### Prepare metadata subset
```{r}
analyze <- metadata %>% 
  select(rowname, id, sample, patient, condition, treatment, annotation, response, group) %>% 
  as_tibble()

analyze %>% head()
```

```{r}
metadata_patient <- metadata %>% 
  select(id, patient, sample, treatment, condition, response, group) %>% 
  unique()

metadata_patient
```

### Global parameters
```{r}
which_annotation <- 'annotation'
```




### Pre vs Mid, Responders
```{r}
group1_filters <- list('condition' = 'PV',
                       'treatment' = 'Pre',
                       'group' = 'Responder')

group2_filters <- list('condition' = 'PV',
                       'treatment' = 'Mid',
                       'group' = 'Responder')

data_subset <- bind_rows(
  analyze %>% apply_filters(group1_filters) %>% mutate(group = 'group1'),
  analyze %>% apply_filters(group2_filters) %>% mutate(group = 'group2')
)


metadata_subset <- data_subset %>% 
  select(id, patient, group) %>% 
  droplevels() %>% 
  unique()
  


  
pseudobulk_subset <- pseudobulk_matrix(seuratobj,
                                       cells = data_subset %>%
                                         filter(annotation == 'Trm') %>%
                                         pull(rowname),
                                       group_by = 'id')

de_il23 <- de_pseudobulk_pipeline(pseudobulk_subset = pseudobulk_subset, 
                                  metadata_subset = metadata_subset,
                                  cluster = 'Trm',
                                  paired = TRUE,
                                  group1_id = 'Responder Pre',
                                  group2_id = 'Responder Mid')



```

### IL17
### Prepare IL17 dataset 
Preprocessed in "secukinumab.Rmd" notebook
#### Load data
```{r}
metadata_il17 <- read_tsv(here('analysis/output/secukinumab/liu_metadata.tsv.gz')) %>% 
  mutate(treatment = 
           case_when(
             Week == 0 ~ 'Pre',
             Week == 12 ~ 'Mid'),
         group = case_when(
           str_detect(PatientOutcome.Week52,'PASI75|PASI100') ~ 'Responder',
           str_detect(PatientOutcome.Week52,'PASI50|limited') ~ 'Non-responder',
         ))

counts_il17 <- read_tsv(here('analysis/output/secukinumab/liu_counts.tsv.gz')) %>% 
  column_to_rownames('gene_id') %>% as.matrix()

ens2symbol <- read_tsv(here('analysis/output/secukinumab/ens2symbol.tsv.gz'))
```
```{r}
metadata_il17
```

```{r}
counts_il17[1:5, 1:5]
```
#### Aggregate matrix by gene name
```{r}
gene_names <- ens2symbol %>% arrange(gene_id, rownames(counts_il17)) %>% pull(gene_name)
counts_il17 <- aggregate_matrix(counts_il17, rows = gene_names)
counts_il17[1:5, 1:5]
```

#### Subjects with PSO at Week 0 and 12
```{r}
metadata_il17_subset <- metadata_il17 %>% 
  filter(QC_Passed,
         treatment %in% c('Pre', 'Mid'),
         Status == 'PSO') %>% 
  select(sample, SubjectId, treatment, CellType, PASI, PASI.Reduction.Week0_Week52, PatientOutcome.Week52, group) %>% 
  group_by(SubjectId) %>% 
  add_count() %>% 
  arrange(SubjectId, treatment, CellType)

metadata_il17_subset
```

#### Run DE
```{r}
de_il17 <- map(c('Teff', 'CD8'), function(i) {
  
  sample_subset <- metadata_il17_subset %>% 
    filter(treatment %in% c('Pre', 'Mid'),
           #group == 'Responder',
           CellType == i) %>% 
    mutate(group = ifelse(treatment == 'Pre', 'group1', 'group2')) %>% 
    select(id = sample, patient = SubjectId, group) %>% 
    add_count(patient) %>% 
    filter(n == 2)
  
  de_il17 <- de_pseudobulk_pipeline(pseudobulk_subset = counts_il17[, sample_subset$id], 
                                    metadata_subset = sample_subset,
                                    cluster = i,
                                    paired = TRUE,
                                    group1_id = 'Responder Pre',
                                    group2_id = 'Responder Mid')
  
  
}) %>% 
  bind_rows() 

de_il17
```

### Combine CD8 and Teff
```{r}
metadata_il17_merge <- metadata_il17_subset %>% 
  filter(CellType %in% c('Teff', 'CD8')) %>% 
  select(sample, SubjectId, treatment, CellType) %>% 
  pivot_wider(names_from = CellType,
              values_from = sample) %>% 
  drop_na() %>% 
  mutate(id = paste0(SubjectId, '_', treatment)) %>% 
  arrange(treatment, id)

metadata_il17_merge
```

```{r}
input <- t(
    counts_il17[,c(metadata_il17_merge$CD8, metadata_il17_merge$Teff)])

rownames(input) <- metadata_il17_merge$id

counts_il17_merge <- aggregate_matrix(input_matrix = input,
                                      rows = rep(metadata_il17_merge$id,2))
```


```{r}
sample_subset <- metadata_il17_merge %>% 
  filter(treatment %in% c('Pre', 'Mid')) %>% 
  mutate(group = ifelse(treatment == 'Pre', 'group1', 'group2')) %>% 
  select(id, patient = SubjectId, group) %>% 
  add_count(patient) %>% 
  filter(n == 2)

de_il17 <- de_pseudobulk_pipeline(pseudobulk_subset = counts_il17_merge[, sample_subset$id], 
                                  metadata_subset = sample_subset,
                                  cluster = 'All',
                                  paired = TRUE,
                                  group1_id = 'IL17 Pre',
                                  group2_id = 'IL17 Mid')
```

# OLD

### Functions
```{r}
aggregate_matrix <- function(input_matrix,
                             rows = NULL) {
  input_matrix <- as.matrix(input_matrix)
  aggregated_matrix <- t(sapply(by(input_matrix, rows, colSums), identity))
}

de_function <- function(counts,
                        metadata,
                        design = c('SubjectId', 'treatment'),
                        reference = 'Mid',
                        ens2symbol = NULL) {
  
  metadata_table <- metadata %>% 
    filter(sample %in% colnames(counts)) %>% 
    droplevels() %>% 
    arrange(sample, colnames(counts))
  
  design <- model.matrix(data = metadata_table, reformulate(design))  
  
  dge <- DGEList(counts = counts,
                 group = metadata_table$treatment)
  
  dge <- dge %>% calcNormFactors(method = 'TMM')
  dge <- dge %>% estimateDisp(design = design)
  fit <- dge %>% glmFit(design = design)
  res <- fit %>% glmLRT()
  
  de_table <- topTags(res, n = Inf) %>% 
    as.data.frame() %>% 
    rownames_to_column('gene') %>% 
    as_tibble() %>% 
    #mutate(logFC = -1 * logFC, # reverse sign such that group 1 is higher
    select(gene, avg_log2FC = logFC, p_val = PValue, padj = FDR, everything())
}

de_function_deseq <- function(counts,
                              metadata,
                              design,
                              reference,
                              ens2symbol = NULL) {
  
  deseq <- DESeqDataSetFromMatrix(countData = counts,
                                  colData = metadata, 
                                  design = reformulate(design))
  
  deseq[[last(design)]] <- relevel(deseq[[last(design)]], ref = reference)
  
  tic()
  deseq <- DESeq(deseq)
  toc()
  
  deseq_res <- results(deseq, tidy = TRUE) %>% 
    as_tibble() %>% 
    mutate(across(where(is.numeric), round, 3))
  
  if(!is.null(ens2symbol)) {
    deseq_res <- inner_join(ens2symbol,
                            deseq_res, by = c('gene_id' = 'row'))
  }
  
  deseq_res
  
}



```


## Analysis
### Prepare IL23 dataset 
#### Load data
```{r}
seuratobj <- read_rds(here(data_dir, 'seuratobj_annotated.rds'))
seuratobj
```

#### Use global or local identities and set as cluster/supercluster in the object
```{r}
which_identity <- 'local'
which_cluster <- paste0(which_identity, '_cluster')
which_supercluster <- paste0(which_identity, '_supercluster')

seuratobj$cluster <- seuratobj[[which_cluster]] # set cluster as main annotation
seuratobj$supercluster <- seuratobj[[which_supercluster]] # set cluster as main annotation
Idents(seuratobj) <- 'cluster'

output_dir <- here(output_dir, which_identity)
dir.create(output_dir)
```

#### Metadata

```{r}
metadata_il23 <- seuratobj@meta.data %>% rownames_to_column()
```

#### IL23 genes table from IL23 profile analysis
```{r}
pv50 <- readxl::read_excel(here('analysis/input/PV50.xlsx'))
pv50_genes <- pv50$...1 %>% intersect(rownames(seuratobj))
```
```{r}
il23_genes_table <- read_tsv(here('analysis/output/09_il23_profile', which_identity, 'il23_genes_table.tsv'))
```
#### CD4 vs CD8 status based on ADT and RNA data
```{r}
cd_data <- FetchData(seuratobj, 
                     vars = c('CD4', 'CD8A', 'CD8B', 'ADT.CD4', 'ADT.CD8', 
                              'cluster', 'supercluster', 'celltype', 'treatment')) %>% 
  rownames_to_column() %>% 
  mutate(ADT = case_when(
    ADT.CD4 == 'Positive' & ADT.CD8 == 'Positive' ~ 'Double-positive',
    ADT.CD4 == 'Positive' ~ 'CD4',
    ADT.CD8 == 'Positive' ~ 'CD8',
    TRUE ~ 'Double-negative'
  ))
```

#### CD8
```{r}
p <- cd_data %>% 
  ggplot(aes(x = CD8A,
             y = CD8B,
             color = ADT)) +
  geom_point(alpha = 0.1,
             size = 0.5) +
  facet_wrap(~ADT)

p
```
#### CD4
```{r}
p <- cd_data %>% 
  ggplot(aes(x = CD4,
             y = CD8A,
             color = ADT)) +
  geom_point(alpha = 0.1,
             size = 0.5) +
  facet_wrap(~ADT)

p
```
#### Treg vs other T cells
```{r}
p <- cd_data %>% 
  filter(celltype == 'Lymphocyte',
         str_detect(supercluster, 'T')) %>% 
  ggplot(aes(x = CD8A,
             y = CD8B,
             color = ADT)) +
  geom_point(alpha = 0.1,
             size = 0.5) +
  facet_wrap(~supercluster)

p
```
#### Classify CD4/8 status based on protein then RNA
```{r}
cd_status <- cd_data %>% 
  mutate(CD = case_when(
    ADT == 'Double-positive' ~ 'Unknown',
    ADT == 'CD4' ~ 'CD4',
    ADT == 'CD8' ~ 'CD8',
    CD4 > 0.01 & (CD8A < 0.01 & CD8B < 0.01) ~ 'CD4',
    (CD8A > 0.01 | CD8B > 0.01) & CD4 < 0.01 ~ 'CD8',
    TRUE ~ 'Unknown'
  ))

cd_status %>% filter(treatment != 'None') %>% pull(CD) %>% table()
```

### Assign each cluster to closest match
```{r}
cluster_cd_tally <- cd_status %>% 
  group_by(cluster, CD) %>% 
  tally() %>% 
  pivot_wider(names_from = CD,
              values_from = n) %>% 
  rowwise() %>% 
  mutate(CD = ifelse(CD4 > CD8, 'CD4', 'CD8'))

cluster_cd_tally 
```

```{r}
cd_status <- cd_status #%>% select(-CD) %>% left_join(cluster_cd_tally %>% select(cluster, CD))

tildra_cd4 <- cd_status %>% filter(CD == 'CD4', !str_detect(supercluster, 'Treg'), str_detect(cluster, 'T'), treatment %in% c('Pre', 'Mid'))
tildra_cd8 <- cd_status %>% filter(CD == 'CD8', !str_detect(supercluster, 'Treg'), str_detect(cluster, 'T'), treatment %in% c('Pre', 'Mid'))

tildra_barcodes <- c(tildra_cd4$rowname, tildra_cd8$rowname)
```

#### Extract counts matrix
```{r}
add_cd <- seuratobj@meta.data %>% rownames_to_column() %>% select(rowname) %>% left_join(cd_status)
seuratobj$CD <- add_cd$CD
counts_il23 <- pseudobulk_matrix(seuratobj,
                                 cells = tildra_barcodes,
                                 group_by = c('id', 'CD'))
counts_il23[1:5, 1:5]
```


```{r}
metadata_il23_subset <- seuratobj@meta.data %>% 
  select(id, CD, patient, treatment, group) %>% 
  unique() %>% 
  filter(CD != 'Unknown',
         treatment != 'None') %>% 
  mutate(sample = paste(id, CD, sep = ':'),
         CellType = ifelse(CD == 'CD4', 'Teff', 'CD8')) %>% 
  as_tibble()
  
metadata_il23_subset
```

#### Run DE
```{r}
de_il23 <- map(c('Teff', 'CD8'), function(i) {
  
  sample_subset <- metadata_il23_subset %>% 
    filter(CellType == i,
           group == 'Responder')
  
  de <- de_function_deseq(counts = as.matrix(counts_il23)[, sample_subset$sample],
                    metadata = sample_subset,
                    design = c('patient', 'treatment'), 
                    reference = 'Mid') %>% 
    mutate(CellType = i)
  
}) %>% 
  bind_rows() 

de_il23
```
```{r}
de_il23 %>% drop_na() %>% filter(padj < 0.05)
```

### Prepare IL23 Trm pseudobulk
```{r}
metadata_trm_subset <- seuratobj@meta.data %>% 
  select(id, patient, treatment, group, supercluster) %>% 
  unique() %>% 
  filter(supercluster == 'Trm',
         treatment != 'None') %>% 
  mutate(sample = paste(id, supercluster, sep = ':')) %>% 
  as_tibble()
  
metadata_trm_subset
```
```{r}
trm_barcodes <- seuratobj@meta.data %>% filter(supercluster == 'Trm', treatment != 'None') %>% rownames()

counts_trm <- pseudobulk_matrix(seuratobj,
                                cells = trm_barcodes,
                                group_by = c('id', 'supercluster'))
```

#### Run DE
```{r}
de_trm <- map(c('Trm'), function(i) {
  
  sample_subset <- metadata_trm_subset %>% 
    filter(supercluster == i,
           group == 'Responder')
  
  de <- de_function(counts = as.matrix(counts_trm)[, sample_subset$sample],
                    metadata = sample_subset,
                    design = c('patient', 'treatment'), 
                    reference = 'Mid') %>% 
    mutate(CellType = i)
  
}) %>% 
  bind_rows() # %>% select(gene, avg_log2FC, p_val, padj) 

de_trm
```
```{r}
de_trm %>% filter(padj < 0.05)
```
### IL23 pseudobulk DE

```{r}
analyze <- seuratobj@meta.data %>% 
  rownames_to_column() %>% 
  select(rowname, id, sample, patient, condition, treatment, annotation = supercluster, response, group) %>% 
  as_tibble()

group1_filters <- list('condition' = 'PV',
                       'treatment' = 'Pre',
                       'group' = 'Responder')

group2_filters <- list('condition' = 'PV',
                       'treatment' = 'Mid',
                       'group' = 'Responder')

data_subset <- bind_rows(
  analyze %>% apply_filters(group1_filters) %>% mutate(group = 'group1'),
  analyze %>% apply_filters(group2_filters) %>% mutate(group = 'group2')
)

pseudobulk_subset <- pseudobulk_matrix(seuratobj,
                                         cells = data_subset %>%
                                           pull(rowname),
                                         group_by = 'id')

sample_subset <- data_subset %>% 
  mutate(sample = id) %>% 
  select(patient, sample, treatment) %>% 
  unique()
  
  de <- de_function(counts = as.matrix(pseudobulk_subset)[, sample_subset$sample],
                    metadata = sample_subset,
                    design = c('patient', 'treatment'), 
                    reference = 'Mid') 
  
    dde_pseudobulk_pipeline(pseudobulk_subset = pseudobulk_subset, 
                         metadata_subset = metadata_subset,
                         cluster = i,
                         paired = TRUE,
                         group1_id = 'Responder Pre',
                         group2_id = 'Responder Mid')
    
```

```{r}
de_il23 <- read_tsv(here('analysis/output/08_de_pseudobulk/local_supercluster/de_pre_vs_mid_responder.tsv.gz')) %>% 
  filter(cluster == 'Trm') 
  
de_il23
```

### Prepare IL17 dataset 
Preprocessed in "secukinumab.Rmd" notebook
#### Load data
```{r}
metadata_il17 <- read_tsv(here('analysis/output/secukinumab/liu_metadata.tsv.gz')) %>% 
  mutate(treatment = 
           case_when(
             Week == 0 ~ 'Pre',
             Week == 12 ~ 'Mid'),
         group = case_when(
           str_detect(PatientOutcome.Week52,'PASI75|PASI100') ~ 'Responder',
           str_detect(PatientOutcome.Week52,'PASI50|limited') ~ 'Non-responder',
         ))

counts_il17 <- read_tsv(here('analysis/output/secukinumab/liu_counts.tsv.gz')) %>% 
  column_to_rownames('gene_id') %>% as.matrix()

ens2symbol <- read_tsv(here('analysis/output/secukinumab/ens2symbol.tsv.gz'))
```
```{r}
metadata_il17
```

```{r}
counts_il17[1:5, 1:5]
```
#### Aggregate matrix by gene name
```{r}
gene_names <- ens2symbol %>% arrange(gene_id, rownames(counts_il17)) %>% pull(gene_name)
counts_il17 <- aggregate_matrix(counts_il17, rows = gene_names)
counts_il17[1:5, 1:5]
```

#### Subjects with PSO at Week 0 and 12
```{r}
metadata_il17_subset <- metadata_il17 %>% 
  filter(QC_Passed,
         treatment %in% c('Pre', 'Mid'),
         Status == 'PSO') %>% 
  select(sample, SubjectId, treatment, CellType, PASI, PASI.Reduction.Week0_Week52, PatientOutcome.Week52, group) %>% 
  group_by(SubjectId) %>% 
  add_count() %>% 
  arrange(SubjectId, treatment, CellType)

metadata_il17_subset
```

#### Run DE
```{r}
de_il17 <- map(c('Teff', 'CD8'), function(i) {
  
  sample_subset <- metadata_il17_subset %>% 
    filter(treatment %in% c('Pre', 'Mid'),
           CellType == i,
           group == 'Responder')
  
  de <- de_function_deseq(counts = counts_il17[, sample_subset$sample],
                    metadata = sample_subset,
                    design = c('SubjectId', 'treatment'),
                    reference = 'Mid') %>% 
    mutate(CellType = i)
  
}) %>% 
  bind_rows() 

de_il17
```
```{r}
de_il17 %>% filter(padj < 0.05)
```

#### Expression table
```{r}
cpm <- t(t(counts_il17) / colSums(counts_il17)) * 1e6
```


```{r}
logcpm <- log10(cpm + 0.01)
```

```{r}
expression_table_il17 <- cpm %>% 
  as.data.frame() %>% 
  rownames_to_column('gene') %>% 
  pivot_longer(cols = -gene,
               names_to = 'sample',
               values_to = 'cpm') %>% 
  inner_join(metadata_il17_subset) %>% 
  mutate(log10cpm = log10(cpm + 1))

expression_table_il17
```
### Compare DE genes
```{r}
de_compare <- inner_join(de_il17 %>% select(gene = row, CellType, log2FoldChange, padj),
                         de_il23 %>% select(gene = row, CellType, log2FoldChange, padj),
                         by = c('gene', 'CellType'),
                         suffix = c('_il17', '_il23'))

de_compare
```
### Genes changed in IL23 and unchanged in IL17

```{r}
genes_of_interest <- de_compare %>% 
  filter(padj_il23 < 0.05, padj_il17 < 0.05, log2FoldChange_il23 > 0) %>% 
  mutate(combination = paste(CellType, gene, sep = ':')) %>% 
  arrange(gene)

genes_of_interest
```


```{r}
radar_genes <- il23_genes_table %>% filter(Status == 'Strict', cluster == 'Trm1') %>% pull(gene)
ray_genes <- c("CBLB", "CHN1", "CXCL13", "DAPK2", "GNLY", "GZMB", "LAG3", "LAYN", "PTPN13", "IL17A", "IL17F", "IL26", "IFNG")

genes_for_plotting <- c(genes_of_interest$gene, radar_genes, ray_genes) %>% unique()

expression_table_il23 <- pseudobulk(seuratobj = seuratobj, 
                                    genes = genes_for_plotting,
                                    cells = tildra_barcodes,
                                    group_by = c('id', 'CD')) %>% 
  left_join(metadata_il23_subset)

expression_table_il17_subset <- expression_table_il17 %>% 
  filter(gene %in% genes_for_plotting,
         CellType %in% c('Teff', 'CD8'))

expression_table_il23 %>% head()
```



```{r}
expression_table_merge <- 
  bind_rows(expression_table_il23 %>% select(gene, patient, treatment, group, CellType, cpm, log10cpm) %>% mutate(therapy = 'Tildra'),
            expression_table_il17_subset %>% select(gene, patient = SubjectId, treatment, group, CellType, cpm, log10cpm) %>% mutate(therapy = 'Secu')) %>% 
  mutate(treatment = factor(treatment, levels = c('Pre', 'Mid')),
         combination = paste(CellType, gene, sep = ':')) 

expression_table_merge
```

#### Plot
```{r}
p <- expression_table_merge %>% 
  filter(combination %in% genes_of_interest$combination) %>% 
  
  ggplot(aes(x = therapy,
             y = log10cpm,
             fill = treatment)) +
  geom_boxplot() +
  facet_wrap(~combination, scales = 'free_y', nrow = 7) +
  theme(legend.position = 'top') +
  scale_fill_few() +
  labs(y = 'log10 counts per million',
       x = 'Biologic Therapy')


ggsave(plot = p,
       h = 10,
       w = 10,
       device = cairo_pdf,
       path = output_dir,
       filename = 'boxplot_all_sig.pdf')

p
```
#### Plot radar plot genes
```{r}
p <- expression_table_merge %>% 
  filter(gene %in% radar_genes) %>% 
  ggplot(aes(x = therapy,
             y = log10cpm,
             fill = treatment)) +
  geom_boxplot() +
  facet_wrap(~combination, scales = 'free_y', nrow = 6) +
  theme(legend.position = 'top') +
  scale_fill_few() +
  labs(y = 'log10 counts per million',
       x = 'Biologic Therapy')

p
```
#### Ray genes
```{r}
p <- expression_table_merge %>% 
  filter(gene %in% ray_genes,
         CellType == 'CD8') %>% 
  ggplot(aes(x = therapy,
             y = log10cpm,
             fill = treatment)) +
  geom_boxplot() +
  facet_wrap(~combination, scales = 'free_y', nrow = 3) +
  theme(legend.position = 'top') +
  scale_fill_few() +
  labs(y = 'log10 counts per million',
       x = 'Biologic Therapy')

p
```
#### Figure: CD8 PV genes
```{r}
cd8_genes <- c(intersect(pv50_genes, 
                         genes_of_interest$gene),'HAVCR2', 
               'IL17A', 'IL17F', 'IL26', 'IFNG') 

gene_stats <- de_compare %>% 
  filter(gene %in% cd8_genes, CellType == 'CD8') %>% 
  mutate(ratio = log2FoldChange_il17/log2FoldChange_il23)
gene_stats
```


```{r}
plot_input <-  expression_table_merge %>% 
  filter(gene %in% cd8_genes,
         CellType == 'CD8') %>% 
  left_join(gene_stats) %>% 
  mutate(gene = factor(gene, levels = cd8_genes)) 

p <- plot_input %>% 
  ggplot(aes(x = therapy,
             y = log10cpm,
             fill = treatment)) +
  geom_boxplot() +
  facet_wrap(~gene, scales = 'free_y', nrow = 2) +
  theme(legend.position = 'top') +
  scale_fill_few() +
  labs(y = 'log10 counts per million',
       x = 'Biologic Therapy')

p

```
### individual plots
```{r}
library(grid)
library(gridExtra)


p_individual <- map(cd8_genes, function(i) {
  
  individual_stats <- plot_input %>% filter(gene == i) %>% select(contains('il'), ratio) %>% unique() %>% 
    mutate(across(everything(), round, 3))
  colnames(individual_stats) <- str_replace(colnames(individual_stats), 'log2FoldChange', 'l2FC')
  
  p <- plot_input %>% 
    filter(gene == i) %>% 
    ggplot(aes(x = therapy,
               y = log10cpm,
               fill = treatment)) +
    geom_boxplot() +
    theme(legend.position = 'top') +
    scale_fill_few() +
    labs(y = 'log10 counts per million',
         x = 'Biologic Therapy') +
    labs(title = i)
  
  p_params <- gridExtra::tableGrob(individual_stats %>% t())
  p_output <- gridExtra::grid.arrange(p, p_params,
                                      #ncol = 2, 
                                      layout_matrix = t(as.matrix(c(1,1,2))))
  gb <- rectGrob(height = 1, width = 1, gp = gpar(lwd = 2, col = "blue", fill = NA)) # border, no fill
  gt <- gTree(children = gList(p_output, gb))
  gt
  
}) %>% patchwork::wrap_plots()
  

```
```{r}
ggsave(plot = p_individual,
       h = 10,
       w = 14,
       device = cairo_pdf,
       path = output_dir,
       filename = 'boxplot_cd8_pv50.pdf')
```

```{r}
intersect(radar_genes, genes_of_interest$gene)
```

```{r}

```

```{r}
intersect(pv50_genes, genes_of_interest$gene)
```
```{r}

```


#### Trm
Venn diagram Overlap of Trm and CD8
```{r}

library(eulerr)
p <- plot(euler(
  list(
    'CD8' = seuratobj@meta.data %>% filter(treatment != 'None', CD == 'CD8') %>% rownames(),
    'CD4' = seuratobj@meta.data %>% filter(treatment != 'None', CD == 'CD4') %>% rownames(),
    'Trm1' = seuratobj@meta.data %>% filter(treatment != 'None', cluster == 'Trm1') %>% rownames())),
    quantities = TRUE,
    fill = c('grey80', 'grey80', 'orangered'), alpha = 0.5)
p
```

```{r}
de_compare_trm <- inner_join(de_il17 %>% select(gene, CellType, log2FoldChange, padj),
                             de_trm %>% select(gene, log2FoldChange, padj),
                             by = c('gene'),
                             suffix = c('_il17', '_trm')) 

de_compare_trm
```
```{r}
genes_of_interest <- de_compare_trm %>% 
  filter(padj_trm < 0.05 & padj_il17 > 0.05 & log2FoldChange_trm > 0) %>% 
  mutate(combination = paste(CellType, gene, sep = ':'))

genes_of_interest
```


```{r}
radar_genes <- il23_genes_table %>% filter(Status == 'Strict', cluster == 'Trm1') %>% pull(gene)

genes_for_plotting <- c(genes_of_interest$gene, radar_genes) %>% unique()

expression_table_trm <- pseudobulk(seuratobj = seuratobj, 
                                    genes = genes_for_plotting,
                                    cells = trm_barcodes,
                                    group_by = c('id', 'cluster')) %>% 
  left_join(metadata_trm_subset)

expression_table_il17_subset <- expression_table_il17 %>% 
  filter(gene %in% genes_for_plotting,
         CellType %in% c('Teff', 'CD8'))

expression_table_il23 %>% head()
```



```{r}
expression_table_merge <- 
  bind_rows(expression_table_trm %>% 
              select(gene, patient, treatment, group, cpm, log10cpm) %>% 
              mutate(therapy = 'Trm',
                     CellType = 'CD8'),
            expression_table_trm %>% 
              select(gene, patient, treatment, group, cpm, log10cpm) %>% 
              mutate(therapy = 'Trm',
                     CellType = 'Teff'),
            expression_table_il17_subset %>% 
              select(gene, patient = SubjectId, treatment, group, CellType, cpm, log10cpm) %>% 
              mutate(therapy = 'Secu')) %>% 
  mutate(treatment = factor(treatment, levels = c('Pre', 'Mid')),
         combination = paste(CellType, gene, sep = ':')) 

expression_table_merge
```

#### Plot
```{r}
p <- expression_table_merge %>% 
  filter(combination %in% genes_of_interest$combination) %>% 
  
  ggplot(aes(x = therapy,
             y = log10cpm,
             fill = treatment)) +
  geom_boxplot() +
  facet_wrap(~combination, scales = 'free_y', nrow = 6) +
  theme(legend.position = 'top') +
  scale_fill_few() +
  labs(y = 'log10 counts per million',
       x = 'Biologic Therapy')

p
```
#### Plot radar plot genes
```{r}
p <- expression_table_merge %>% 
  filter(gene %in% radar_genes) %>% 
  ggplot(aes(x = therapy,
             y = log10cpm,
             fill = treatment)) +
  geom_boxplot() +
  facet_wrap(~combination, scales = 'free_y', nrow = 6) +
  theme(legend.position = 'top') +
  scale_fill_few() +
  labs(y = 'log10 counts per million',
       x = 'Biologic Therapy')

p
```
```{r}
ggsave(plot = p,
       h = 10,
       w = 10,
       device = cairo_pdf,
       path = output_dir,
       filename = 'boxplot_trm_sig.pdf')
```



### Export tables of interest
```{r}
de_compare %>% mutate(significant_il23 = ifelse(padj_il23 < 0.05, TRUE, FALSE),
                      significant_il17 = ifelse(padj_il17 < 0.05, TRUE, FALSE),
                      radar_gene = ifelse(gene %in% radar_genes, TRUE, FALSE),
                      pv50_gene = ifelse(gene %in% pv50_genes, TRUE, FALSE)) %>% 
  write_tsv(here(output_dir, 'deg_comparison.tsv'))
```
```{r}
de_compare_trm %>% mutate(significant_trm = ifelse(padj_trm < 0.05, TRUE, FALSE),
                          significant_il17 = ifelse(padj_il17 < 0.05, TRUE, FALSE),
                          radar_gene = ifelse(gene %in% radar_genes, TRUE, FALSE),
                          pv50_gene = ifelse(gene %in% pv50_genes, TRUE, FALSE)) %>% 
  write_tsv(here(output_dir, 'deg_comparison_trm.tsv'))
```



## Session Info
```{r}
sessionInfo()
```

### OLD
```{r}

de_il23_group_prepare <- de_pre_vs_mid_response %>% 
                               filter(cluster == 'Trm1') %>% 
                               select(gene, avg_log2FC_Trm1 = avg_log2FC, padj_Trm1 = padj)

de_il23_patient_prepare <- de_pre_vs_mid_patients %>% 
  filter(cluster == 'Trm1') %>% 
  select(gene, patient, avg_log2FC, padj) %>% 
  pivot_wider(names_from = 'patient',
              values_from = c('avg_log2FC', 'padj'))
```



### Comparison with Trms
```{r}
de_il17_prepare <- de_il17 %>% 
  mutate(log2FoldChange = -log2FoldChange) %>% 
  select(CellType,
         gene = gene,
         log2FoldChange,
         padj) %>% 
  group_by(CellType, gene) %>% 
  slice_max(abs(log2FoldChange)) %>% 
  slice_min(padj) %>% 
  ungroup() %>% 
  pivot_wider(names_from = 'CellType',
              values_from = c('log2FoldChange', 'padj'))

de_il23_prepare <- de_il23
```


```{r}
de_trm_prepare <- de_trm

de_compare_trm <- inner_join(de_il17_prepare,
                             de_il23_group_prepare) %>% 
  inner_join(de_il23_patient_prepare) %>% 

de_compare_trm %>% write_tsv(here(output_dir, 'trm_compare.tsv'))

```
```{r}

```
