==> analysis/scripts/13_trajectory.Rmd <==
---
title: "Trajectory analysis"
author: "David Wu"
---

Set working directory to project directory
```{r setup}
library(here)
```

### Optionally convert this Rmd to R script 
```{r eval=FALSE, message=FALSE}
current_file <- rstudioapi::getActiveDocumentContext()$path
output_file <- stringr::str_replace(current_file, '.Rmd', '.R')
knitr::purl(current_file, output = output_file)
file.edit(output_file)
```

### Directories
```{r, message=FALSE}
output_dir <- here('analysis/output/13_trajectory') # analysis file output directory
data_dir <- here('data/derived/tildra') # data file output directory

dir.create(output_dir, showWarnings = FALSE)
dir.create(data_dir, showWarnings = FALSE)
```

### Loading the libraries
```{r, message=FALSE}
library(Seurat) 
library(tidyverse)
library(ggExtra)
library(patchwork)
library(ggthemes)
library(ggforce)
library(wutilities) # devtools::install_github('symbiologist/wutilities')
library(tictoc)
library(rcartocolor)
library(ggrepel)
library(harmony)
theme_set(theme_dwu()) # set default theme
```
#### Custom functions
```{r}
recluster <- function(seuratobj, 
                      assay = 'RNA',
                      reduction = 'harmony',
                      dims = 1:20,
                      seed = 42,
                      n.neighbors = 30,
                      n.epochs = 200,
                      min.dist = 0.3,
                      k.param = n.neighbors,
                      metric = 'cosine',
                      algorithm = 1,
                      resolution = c(0.1, 1.0, 2.0),
                      default_ident = 'RNA_snn_res.1') {
  
    ### RunUMAP, FindNeighbors, and FindClusters 
  
  seuratobj <- seuratobj %>% 
    FindNeighbors(reduction = reduction, 
                  dims = dims, 
                  assay = assay,
                  annoy.metric = metric,
                  k.param = k.param) %>% 
    FindClusters(random.seed = seed,
                 algorithm = algorithm,
                 resolution = resolution) %>% 
    RunUMAP(reduction = reduction, 
            dims = dims, 
            assay = assay, 
            seed.use = seed,
            n.neighbors = n.neighbors,
            n.epochs = n.epochs,
            min.dist = min.dist)
  
  ### Fix cluster factor levels
  clusterings <- colnames(seuratobj@meta.data) %>% str_subset('_res')
  
  for(i in clusterings) {
    clusters <- seuratobj@meta.data[[i]]
    seuratobj[[i]] <- factor(clusters, levels = levels(clusters) %>% as.numeric() %>% sort())
  }
  
  Idents(seuratobj) <- default_ident
  seuratobj$seurat_clusters <- seuratobj[[default_ident]]
  
  seuratobj
}
```

#### Load data
```{r}
seuratobj <- read_rds(here(data_dir, 'seuratobj_annotated.rds'))
seuratobj
```
### Use global or local identities and set as cluster/supercluster in the object
```{r}
which_identity <- 'local'
which_level <- 'cluster'
which_annotation <- paste0(which_identity, '_', which_level)

which_cluster <- paste0(which_identity, '_cluster')
which_supercluster <- paste0(which_identity, '_supercluster')
which_ultracluster <- paste0(which_identity, '_ultracluster')

seuratobj$cluster <- seuratobj[[which_cluster]] 
seuratobj$supercluster <- seuratobj[[which_supercluster]] 
seuratobj$ultracluster <- seuratobj[[which_ultracluster]] 
seuratobj$annotation <- seuratobj[[which_annotation]]
Idents(seuratobj) <- 'annotation'
```
#### Subset TRMs
```{r}
which_celltype <- seuratobj$annotation %>% str_subset('Trm') %>% unique() %>% sort() %>% head(1)
subsetobj <- seuratobj %>% subset(annotation == which_celltype & condition != 'AD' & group %in% c(NA, 'Responder'))

subsetobj$description <- paste0(subsetobj$condition, subsetobj$group, subsetobj$treatment) %>% 
  str_replace('HCNANone', 'HC') %>% 
  str_remove('PV') %>% 
  str_replace('NANone', 'PV') %>% 
  factor(levels = c('PV',
                    'HC',
                    # 'Non-responderPre',
                    # 'Non-responderMid',
                    'ResponderPre',
                    'ResponderMid'))
  
metadata <- subsetobj@meta.data %>% rownames_to_column()

subsetobj
```
#### Prepare
```{r}
subsetobj <- subsetobj %>% 
  NormalizeData() %>% 
  FindVariableFeatures() %>% 
  ScaleData() %>% 
  RunPCA()
```
```{r}
ElbowPlot(subsetobj, ndims = 50)
```

### Reharmonize
```{r}
batch <- c('flow', 'mouse', 'chemistry') # global batch variables
max_harmony_dims <- 20
subsetobj <- subsetobj %>% 
  RunHarmony(group.by.vars = batch,
             dims.use = 1:max_harmony_dims)
```
```{r}
p <- subsetobj %>% seurat_feature(features = 'description', facet_hide = TRUE, reduction = 'harmony')
p
```
```{r}
p <- subsetobj %>% seurat_feature(features = 'description', reduction = 'pca')
p
```
### Recluster
```{r}
max_dims <- 20
subsetobj <- subsetobj %>% recluster(dims = 1:max_dims)
```

```{r}
p <- subsetobj %>% 
  seurat_feature(features = 'description', 
                 facet_hide = TRUE,
                 alpha = 0.2, 
                 size = 0.5)
p
```
```{r}
p <- subsetobj %>% seurat_feature(features = 'description', 
                                  facets = 'description',
                                  label = FALSE,
                                  legend_position = 'none',
                                  alpha = 0.5, 
                                  size = 0.5)
p
```
```{r}
p <- subsetobj %>% seurat_feature(features = 'description', 
                                  facets = 'description',
                                  label = FALSE,
                                  legend_position = 'none',
                                  nrow = 1,
                                  alpha = 0.2, 
                                  size = 0.5,
                                  reduction = 'pca',
                                  dims = 2:1) +
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_vline(xintercept = 0, linetype = 'dashed')

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 2.5,
       w = 8,
	filename = 'trajectory_facets_pca.pdf')
p
```

```{r}
subsetobj <- subsetobj %>% 
  FindNeighbors() %>% 
  FindClusters(resolutio = c(0.2, 0.4, 0.8))


```

```{r}
subsetobj %>% seurat_feature(features = 'seurat_clusters', 
                                  facets = 'description',
                                  label = TRUE,
                                  legend_position = 'none',
                                  nrow = 1,
                                  alpha = 0.2, 
                                  size = 0.5,
                                  reduction = 'pca',
                                  dims = 2:1) +
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_vline(xintercept = 0, linetype = 'dashed')
```
```{r}
subsetobj %>% seurat_feature(features = 'RNA_snn_res.0.8', 
                                  label = TRUE,
                                  legend_position = 'none',
                                  nrow = 1,
                                  alpha = 0.2, 
                                  size = 0.5,
                                  reduction = 'pca',
                                  dims = 2:1) +
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_vline(xintercept = 0, linetype = 'dashed')
```
```{r}
markers_hc <- FindMarkers(subsetobj, ident.1 = 0) %>% 
  rownames_to_column('gene') %>% 
  mutate(pct.diff = pct.1 - pct.2) %>%
  arrange(-pct.diff)
```

```{r}
markers_hc 
```

```{r}
markers_hc %>% write_tsv(here(output_dir, 'markers_hc.tsv'))
```

```{r}
seurat_feature(subsetobj, features = c('TXNIP', 'PRDM1', 'CCR4', 'LUZP1'), reduction = 'pca', dims = 2:1, size = 0.5)
```


### Resume
```{r}
p <- subsetobj %>% seurat_feature(features = 'description', 
                                  facets = 'description',
                                  label = FALSE,
                                  legend_position = 'none',
                                  nrow = 1,
                                  alpha = 0.2, 
                                  size = 0.5,
                                  reduction = 'harmony',
                                  dims = 2:1) +
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_vline(xintercept = 0, linetype = 'dashed')

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 2.5,
       w = 8,
	filename = 'trajectory_facets_harmony.pdf')
p
```
```{r}
p <- subsetobj %>% seurat_feature(features = 'description', 
                                  facets = 'description',
                                  label = FALSE,
                                  legend_position = 'none',
                                  nrow = 1,
                                  alpha = 0.2, 
                                  size = 0.5,
                                  reduction = 'harmony',
                                  dims = 1:2) +
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_vline(xintercept = 0, linetype = 'dashed')

ggsave(plot = p,
       device = cairo_pdf,
       path = output_dir,
       h = 2,
       w = 10,
	filename = 'trajectory_facets_harmony.pdf')
p
```

```{r}
p <- subsetobj %>% seurat_feature(features = 'description', 
                                  label = TRUE,
                                  alpha = 0.5, 
                                  size = 0.5,
                                  facet_hide = TRUE,
                                  reduction = 'pca')
p
```
```{r}
p <- map(subsetobj$description %>% levels(), function(i) {
  
  density_plot(subsetobj, 
               x = 'PC_1', 
               y = 'PC_2', 
               group = 'description', 
               target = i, 
               point_alpha = 0.5, 
               point_size = 0.5,
               point_color = 'grey80') +
    ggtitle(i)
  
}) %>% wrap_plots(nrow = 1)

p

```

```{r}
p <- map(subsetobj$description %>% levels(), function(i) {
  
  density_plot(subsetobj, 
               x = 'harmony_1', 
               y = 'harmony_2', 
               group = 'description', 
               target = i, 
               point_alpha = 0.5, 
               point_size = 0.5,
               point_color = 'grey80') +
    ggtitle(i)
  
}) %>% wrap_plots(nrow = 1)

p
```

```{r}
p <- map(subsetobj$description %>% levels(), function(i) {
  
  density_plot(subsetobj, 
               x = 'UMAP_1', 
               y = 'UMAP_2', 
               group = 'description', 
               target = i, 
               point_alpha = 0.5, 
               point_size = 0.5,
               point_color = 'grey80') +
    ggtitle(i)
  
}) %>% wrap_plots(nrow = 1)

p
```

```{r}
sessionInfo()
```
